Naive Bayes Classification Learner

Naive Bayes classification. Calls e1071::naiveBayes() from package e1071.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.naive_bayes")
lrn("classif.naive_bayes")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3learners, e1071

Parameters

Id	Type	Default	Range

eps	numeric	0	\((-\infty, \infty)\)

laplace	numeric	0	\([0, \infty)\)

threshold	numeric	0.001	\((-\infty, \infty)\)

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayes

Methods

Public methods

LearnerClassifNaiveBayes$new()

LearnerClassifNaiveBayes$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("e1071", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("classif.naive_bayes")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerClassifNaiveBayes:classif.naive_bayes>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, e1071
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor
#> * Properties: multiclass, twoclass
#> 
#> Naive Bayes Classifier for Discrete Predictors
#> 
#> Call:
#> naiveBayes.default(x = x, y = y)
#> 
#> A-priori probabilities:
#> y
#>         M         R 
#> 0.5323741 0.4676259 
#> 
#> Conditional probabilities:
#>    V1
#> y         [,1]       [,2]
#>   M 0.03343108 0.02739813
#>   R 0.02244923 0.01521737
#> 
#>    V10
#> y        [,1]       [,2]
#>   M 0.2460000 0.12987396
#>   R 0.1528985 0.09439155
#> 
#>    V11
#> y        [,1]      [,2]
#>   M 0.2857459 0.1166163
#>   R 0.1722600 0.1012884
#> 
#>    V12
#> y        [,1]      [,2]
#>   M 0.2998689 0.1273317
#>   R 0.1971415 0.1236892
#> 
#>    V13
#> y        [,1]      [,2]
#>   M 0.3147986 0.1416027
#>   R 0.2376538 0.1378416
#> 
#>    V14
#> y        [,1]      [,2]
#>   M 0.3204216 0.1732725
#>   R 0.2761492 0.1765361
#> 
#>    V15
#> y        [,1]      [,2]
#>   M 0.3368149 0.2023379
#>   R 0.3070431 0.2328125
#> 
#>    V16
#> y        [,1]      [,2]
#>   M 0.3870595 0.2185413
#>   R 0.3798615 0.2661840
#> 
#>    V17
#> y        [,1]      [,2]
#>   M 0.4196608 0.2548757
#>   R 0.4139138 0.2955339
#> 
#>    V18
#> y        [,1]      [,2]
#>   M 0.4592878 0.2728964
#>   R 0.4393246 0.2661027
#> 
#>    V19
#> y        [,1]      [,2]
#>   M 0.5314851 0.2706937
#>   R 0.4509046 0.2496286
#> 
#>    V2
#> y         [,1]       [,2]
#>   M 0.04270946 0.03568013
#>   R 0.02962769 0.02502134
#> 
#>    V20
#> y       [,1]      [,2]
#>   M 0.612273 0.2672022
#>   R 0.484980 0.2579316
#> 
#>    V21
#> y        [,1]      [,2]
#>   M 0.6685311 0.2600359
#>   R 0.5349477 0.2357076
#> 
#>    V22
#> y        [,1]      [,2]
#>   M 0.6766284 0.2459926
#>   R 0.5495108 0.2554554
#> 
#>    V23
#> y        [,1]      [,2]
#>   M 0.6785757 0.2642858
#>   R 0.6017046 0.2454930
#> 
#>    V24
#> y        [,1]      [,2]
#>   M 0.6818932 0.2497661
#>   R 0.6568862 0.2356681
#> 
#>    V25
#> y        [,1]      [,2]
#>   M 0.6674419 0.2610022
#>   R 0.6772123 0.2667411
#> 
#>    V26
#> y        [,1]      [,2]
#>   M 0.6954122 0.2559616
#>   R 0.7000015 0.2410434
#> 
#>    V27
#> y        [,1]      [,2]
#>   M 0.6989041 0.2856959
#>   R 0.6918615 0.2178490
#> 
#>    V28
#> y        [,1]      [,2]
#>   M 0.6978216 0.2677868
#>   R 0.6863062 0.2140017
#> 
#>    V29
#> y        [,1]      [,2]
#>   M 0.6360378 0.2439305
#>   R 0.6522892 0.2279167
#> 
#>    V3
#> y         [,1]       [,2]
#>   M 0.04991892 0.04123507
#>   R 0.03338000 0.02688670
#> 
#>    V30
#> y        [,1]      [,2]
#>   M 0.5761270 0.2108829
#>   R 0.5817062 0.2374751
#> 
#>    V31
#> y        [,1]      [,2]
#>   M 0.4798541 0.2285096
#>   R 0.5317108 0.2005047
#> 
#>    V32
#> y        [,1]      [,2]
#>   M 0.4367635 0.2186505
#>   R 0.4511431 0.1999408
#> 
#>    V33
#> y        [,1]      [,2]
#>   M 0.4063432 0.2087250
#>   R 0.4455031 0.2124249
#> 
#>    V34
#> y        [,1]      [,2]
#>   M 0.3766757 0.2175144
#>   R 0.4488431 0.2363194
#> 
#>    V35
#> y        [,1]      [,2]
#>   M 0.3542419 0.2642669
#>   R 0.4582031 0.2601686
#> 
#>    V36
#> y       [,1]      [,2]
#>   M 0.343173 0.2678238
#>   R 0.471440 0.2635036
#> 
#>    V37
#> y        [,1]      [,2]
#>   M 0.3317797 0.2456418
#>   R 0.4165092 0.2452177
#> 
#>    V38
#> y        [,1]      [,2]
#>   M 0.3399257 0.2149889
#>   R 0.3466215 0.2158030
#> 
#>    V39
#> y        [,1]      [,2]
#>   M 0.3327892 0.1876061
#>   R 0.3150815 0.2097415
#> 
#>    V4
#> y         [,1]       [,2]
#>   M 0.06490405 0.04385078
#>   R 0.03739846 0.02556798
#> 
#>    V40
#> y        [,1]      [,2]
#>   M 0.2998730 0.1662433
#>   R 0.3115923 0.1995398
#> 
#>    V41
#> y        [,1]      [,2]
#>   M 0.2893243 0.1758636
#>   R 0.2774308 0.1826720
#> 
#>    V42
#> y        [,1]      [,2]
#>   M 0.2891743 0.1627690
#>   R 0.2610523 0.1749285
#> 
#>    V43
#> y        [,1]      [,2]
#>   M 0.2652676 0.1308424
#>   R 0.2242215 0.1371821
#> 
#>    V44
#> y        [,1]      [,2]
#>   M 0.2429311 0.1452064
#>   R 0.1759031 0.1161148
#> 
#>    V45
#> y        [,1]       [,2]
#>   M 0.2434270 0.17691314
#>   R 0.1415215 0.09734069
#> 
#>    V46
#> y        [,1]       [,2]
#>   M 0.1980392 0.13826533
#>   R 0.1205138 0.09395306
#> 
#>    V47
#> y         [,1]       [,2]
#>   M 0.14501081 0.08414439
#>   R 0.09404923 0.07185234
#> 
#>    V48
#> y         [,1]       [,2]
#>   M 0.10561351 0.06343938
#>   R 0.07061846 0.05247121
#> 
#>    V49
#> y         [,1]       [,2]
#>   M 0.06137703 0.03478858
#>   R 0.03892615 0.03405606
#> 
#>    V5
#> y         [,1]       [,2]
#>   M 0.08505270 0.05199392
#>   R 0.05997846 0.04270175
#> 
#>    V50
#> y         [,1]       [,2]
#>   M 0.02261486 0.01230729
#>   R 0.01867231 0.01364706
#> 
#>    V51
#> y         [,1]        [,2]
#>   M 0.01793649 0.010283203
#>   R 0.01192923 0.008419745
#> 
#>    V52
#> y         [,1]        [,2]
#>   M 0.01565000 0.010272883
#>   R 0.01049692 0.007209563
#> 
#>    V53
#> y          [,1]        [,2]
#>   M 0.010944595 0.006795811
#>   R 0.009847692 0.005592967
#> 
#>    V54
#> y          [,1]        [,2]
#>   M 0.012420270 0.008938572
#>   R 0.009692308 0.005632226
#> 
#>    V55
#> y          [,1]        [,2]
#>   M 0.010517568 0.008525953
#>   R 0.008523077 0.004846254
#> 
#>    V56
#> y          [,1]        [,2]
#>   M 0.008835135 0.005991374
#>   R 0.007593846 0.004699131
#> 
#>    V57
#> y          [,1]        [,2]
#>   M 0.007931081 0.005798888
#>   R 0.007270769 0.004857813
#> 
#>    V58
#> y          [,1]        [,2]
#>   M 0.009564865 0.007251328
#>   R 0.006096923 0.004419346
#> 
#>    V59
#> y          [,1]        [,2]
#>   M 0.009229730 0.007247668
#>   R 0.007106154 0.005198614
#> 
#>    V6
#> y         [,1]       [,2]
#>   M 0.11350946 0.05303901
#>   R 0.09595231 0.06248806
#> 
#>    V60
#> y          [,1]        [,2]
#>   M 0.006960811 0.006198582
#>   R 0.005969231 0.003454703
#> 
#>    V7
#> y        [,1]       [,2]
#>   M 0.1313405 0.06026026
#>   R 0.1122615 0.06268613
#> 
#>    V8
#> y        [,1]       [,2]
#>   M 0.1568703 0.09546334
#>   R 0.1141554 0.07723990
#> 
#>    V9
#> y        [,1]       [,2]
#>   M 0.2151595 0.12987358
#>   R 0.1315492 0.08467338
#> 
#> classif.ce 
#>  0.3478261