Classification Neural Network Learner

Single Layer Neural Network. Calls nnet::nnet.formula() from package nnet.

Note that modern neural networks with multiple layers are connected via package mlr3keras.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.nnet")
lrn("classif.nnet")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3learners, nnet

Parameters

Id	Type	Default	Levels	Range

Hess	logical	FALSE	TRUE, FALSE	-

MaxNWts	integer	1000		\([1, \infty)\)

Wts	untyped	-		-

abstol	numeric	1e-04		\((-\infty, \infty)\)

censored	logical	FALSE	TRUE, FALSE	-

contrasts	untyped			-

decay	numeric	0		\((-\infty, \infty)\)

mask	untyped	-		-

maxit	integer	100		\([1, \infty)\)

na.action	untyped	-		-

rang	numeric	0.7		\((-\infty, \infty)\)

reltol	numeric	1e-08		\((-\infty, \infty)\)

size	integer	3		\([0, \infty)\)

skip	logical	FALSE	TRUE, FALSE	-

subset	untyped	-		-

trace	logical	TRUE	TRUE, FALSE	-

formula	untyped	-		-

Initial parameter values

size:

Adjusted default: 3L.

Reason for change: no default in nnet().

Custom mlr3 parameters

formula: if not provided, the formula is set to task$formula().

References

Ripley BD (1996). Pattern Recognition and Neural Networks. Cambridge University Press. doi:10.1017/cbo9780511812651 .

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNnet

Methods

Public methods

LearnerClassifNnet$new()

LearnerClassifNnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNnet$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("nnet", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("classif.nnet")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerClassifNnet:classif.nnet>: Single Layer Neural Network
#> * Model: -
#> * Parameters: size=3
#> * Packages: mlr3, mlr3learners, nnet
#> * Predict Types:  response, [prob]
#> * Feature Types: integer, numeric, factor, ordered
#> * Properties: multiclass, twoclass, weights
#> # weights:  187
#> initial  value 98.061119 
#> iter  10 value 62.681638
#> iter  20 value 30.474401
#> iter  30 value 16.002467
#> iter  40 value 15.785337
#> iter  50 value 15.777706
#> iter  60 value 15.777335
#> final  value 15.777311 
#> converged
#> a 60-3-1 network with 187 weights
#> inputs: V1 V10 V11 V12 V13 V14 V15 V16 V17 V18 V19 V2 V20 V21 V22 V23 V24 V25 V26 V27 V28 V29 V3 V30 V31 V32 V33 V34 V35 V36 V37 V38 V39 V4 V40 V41 V42 V43 V44 V45 V46 V47 V48 V49 V5 V50 V51 V52 V53 V54 V55 V56 V57 V58 V59 V6 V60 V7 V8 V9 
#> output(s): Class 
#> options were - entropy fitting 
#> classif.ce 
#>  0.1449275