Kriging Regression Learner

Kriging regression. Calls DiceKriging::km() from package DiceKriging.

The predict type hyperparameter "type" defaults to "sk" (simple kriging).

The additional hyperparameter nugget.stability is used to overwrite the hyperparameter nugget with nugget.stability * var(y) before training to improve the numerical stability. We recommend a value of 1e-8.

The additional hyperparameter jitter can be set to add N(0, [jitter])-distributed noise to the data before prediction to avoid perfect interpolation. We recommend a value of 1e-12.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.km")
lrn("regr.km")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3learners, DiceKriging

Parameters

Id	Type	Default	Levels	Range

bias.correct	logical	FALSE	TRUE, FALSE	-

checkNames	logical	TRUE	TRUE, FALSE	-

coef.cov	untyped			-

coef.trend	untyped			-

coef.var	untyped			-

control	untyped			-

cov.compute	logical	TRUE	TRUE, FALSE	-

covtype	character	matern5_2	gauss, matern5_2, matern3_2, exp, powexp	-

estim.method	character	MLE	MLE, LOO	-

gr	logical	TRUE	TRUE, FALSE	-

iso	logical	FALSE	TRUE, FALSE	-

jitter	numeric	0		\([0, \infty)\)

kernel	untyped			-

knots	untyped			-

light.return	logical	FALSE	TRUE, FALSE	-

lower	untyped			-

multistart	integer	1		\((-\infty, \infty)\)

noise.var	untyped			-

nugget	numeric	-		\((-\infty, \infty)\)

nugget.estim	logical	FALSE	TRUE, FALSE	-

nugget.stability	numeric	0		\([0, \infty)\)

optim.method	character	BFGS	BFGS, gen	-

parinit	untyped			-

penalty	untyped			-

scaling	logical	FALSE	TRUE, FALSE	-

se.compute	logical	TRUE	TRUE, FALSE	-

type	character	SK	SK, UK	-

upper	untyped			-

References

Roustant O, Ginsbourger D, Deville Y (2012). “DiceKriging, DiceOptim: Two R Packages for the Analysis of Computer Experiments by Kriging-Based Metamodeling and Optimization.” Journal of Statistical Software, 51(1), 1--55. doi:10.18637/jss.v051.i01 .

See also

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrKM

Methods

Public methods

LearnerRegrKM$new()

LearnerRegrKM$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrKM$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrKM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("DiceKriging", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("regr.km")
print(learner)

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerRegrKM:regr.km>: Kriging
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, DiceKriging
#> * Predict Types:  [response], se
#> * Feature Types: logical, integer, numeric
#> * Properties: -
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : BFGS 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern5_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 
#>   - parameters upper bounds :  2 6 8 786.6 3.34 4 404 16.8 2 7.474 
#>   - best initial criterion value(s) :  -61.64327 
#> 
#> N = 10, M = 5 machine precision = 2.22045e-16
#> At X0, 0 variables are exactly at the bounds
#> At iterate     0  f=       61.643  |proj g|=       3.3851
#> At iterate     1  f =       61.459  |proj g|=        0.6139
#> At iterate     2  f =       61.434  |proj g|=       0.45356
#> At iterate     3  f =        61.39  |proj g|=       0.43257
#> At iterate     4  f =       61.349  |proj g|=       0.95242
#> At iterate     5  f =       61.223  |proj g|=        2.2176
#> At iterate     6  f =       61.115  |proj g|=        2.2429
#> At iterate     7  f =       60.985  |proj g|=       0.19911
#> At iterate     8  f =       60.928  |proj g|=       0.32015
#> ys=-6.958e-03  -gs= 5.441e-02, BFGS update SKIPPED
#> At iterate     9  f =       60.817  |proj g|=       0.60774
#> At iterate    10  f =       60.802  |proj g|=        0.6172
#> At iterate    11  f =       60.749  |proj g|=        0.6478
#> At iterate    12  f =       60.703  |proj g|=       0.63978
#> At iterate    13  f =       60.608  |proj g|=       0.20062
#> At iterate    14  f =         60.6  |proj g|=       0.15816
#> At iterate    15  f =       60.595  |proj g|=      0.064078
#> At iterate    16  f =       60.586  |proj g|=       0.11644
#> At iterate    17  f =       60.574  |proj g|=       0.28153
#> At iterate    18  f =       60.554  |proj g|=       0.26886
#> At iterate    19  f =       60.553  |proj g|=       0.12288
#> At iterate    20  f =       60.548  |proj g|=       0.10517
#> At iterate    21  f =       60.542  |proj g|=        0.2464
#> At iterate    22  f =       60.525  |proj g|=       0.46878
#> At iterate    23  f =       60.501  |proj g|=       0.73387
#> At iterate    24  f =       60.479  |proj g|=       0.60182
#> At iterate    25  f =       60.456  |proj g|=       0.07737
#> At iterate    26  f =       60.455  |proj g|=      0.057897
#> At iterate    27  f =       60.453  |proj g|=      0.059506
#> At iterate    28  f =       60.445  |proj g|=      0.062301
#> At iterate    29  f =       60.438  |proj g|=      0.074616
#> At iterate    30  f =       60.428  |proj g|=      0.083749
#> At iterate    31  f =       60.417  |proj g|=       0.48098
#> At iterate    32  f =       60.403  |proj g|=       0.13275
#> At iterate    33  f =       60.395  |proj g|=      0.058783
#> At iterate    34  f =       60.392  |proj g|=      0.015994
#> At iterate    35  f =       60.392  |proj g|=     0.0058649
#> At iterate    36  f =       60.392  |proj g|=     0.0058878
#> At iterate    37  f =       60.392  |proj g|=      0.011264
#> At iterate    38  f =       60.392  |proj g|=      0.018069
#> At iterate    39  f =       60.392  |proj g|=       0.10744
#> At iterate    40  f =       60.391  |proj g|=      0.096627
#> At iterate    41  f =       60.387  |proj g|=       0.10352
#> At iterate    42  f =       60.385  |proj g|=      0.087352
#> At iterate    43  f =       60.384  |proj g|=      0.064432
#> At iterate    44  f =       60.383  |proj g|=      0.021176
#> At iterate    45  f =       60.383  |proj g|=      0.049358
#> At iterate    46  f =       60.383  |proj g|=      0.030606
#> At iterate    47  f =       60.383  |proj g|=      0.007431
#> At iterate    48  f =       60.383  |proj g|=     0.0056912
#> At iterate    49  f =       60.383  |proj g|=     0.0053842
#> At iterate    50  f =       60.383  |proj g|=      0.002153
#> At iterate    51  f =       60.383  |proj g|=     0.0059569
#> At iterate    52  f =       60.383  |proj g|=      0.011505
#> At iterate    53  f =       60.383  |proj g|=      0.016452
#> At iterate    54  f =       60.383  |proj g|=       0.04904
#> At iterate    55  f =       60.383  |proj g|=      0.027198
#> At iterate    56  f =       60.383  |proj g|=      0.002892
#> At iterate    57  f =       60.383  |proj g|=     0.0076983
#> At iterate    58  f =       60.383  |proj g|=      0.013942
#> At iterate    59  f =       60.382  |proj g|=      0.030731
#> At iterate    60  f =       60.382  |proj g|=      0.039102
#> At iterate    61  f =       60.375  |proj g|=       0.26159
#> At iterate    62  f =       60.375  |proj g|=       0.26557
#> At iterate    63  f =       60.369  |proj g|=       0.14335
#> At iterate    64  f =       60.349  |proj g|=       0.14735
#> At iterate    65  f =       60.344  |proj g|=       0.26393
#> At iterate    66  f =       60.342  |proj g|=       0.13406
#> At iterate    67  f =       60.342  |proj g|=      0.074819
#> At iterate    68  f =       60.341  |proj g|=      0.076339
#> At iterate    69  f =        60.34  |proj g|=      0.071832
#> At iterate    70  f =       60.339  |proj g|=      0.072416
#> At iterate    71  f =       60.329  |proj g|=       0.12871
#> At iterate    72  f =       60.319  |proj g|=       0.21699
#> At iterate    73  f =       60.317  |proj g|=      0.024963
#> At iterate    74  f =       60.316  |proj g|=      0.026209
#> At iterate    75  f =       60.315  |proj g|=      0.024936
#> At iterate    76  f =       60.314  |proj g|=       0.19685
#> At iterate    77  f =       60.313  |proj g|=      0.014005
#> At iterate    78  f =       60.312  |proj g|=      0.016355
#> At iterate    79  f =       60.312  |proj g|=     0.0063627
#> At iterate    80  f =       60.312  |proj g|=     0.0023268
#> At iterate    81  f =       60.312  |proj g|=      0.002268
#> At iterate    82  f =       60.312  |proj g|=     0.0099927
#> At iterate    83  f =       60.312  |proj g|=      0.005813
#> At iterate    84  f =       60.312  |proj g|=      0.012544
#> At iterate    85  f =       60.311  |proj g|=      0.039386
#> At iterate    86  f =       60.311  |proj g|=       0.08133
#> At iterate    87  f =        60.31  |proj g|=      0.085598
#> At iterate    88  f =        60.31  |proj g|=      0.061953
#> At iterate    89  f =        60.31  |proj g|=       0.05304
#> At iterate    90  f =        60.31  |proj g|=     0.0054505
#> At iterate    91  f =        60.31  |proj g|=     0.0015457
#> At iterate    92  f =        60.31  |proj g|=     0.0032966
#> At iterate    93  f =        60.31  |proj g|=     0.0015382
#> At iterate    94  f =        60.31  |proj g|=      0.036468
#> At iterate    95  f =        60.31  |proj g|=     0.0085287
#> At iterate    96  f =        60.31  |proj g|=     0.0063556
#> At iterate    97  f =        60.31  |proj g|=     0.0050037
#> At iterate    98  f =        60.31  |proj g|=     0.0075682
#> At iterate    99  f =       60.309  |proj g|=     0.0091391
#> At iterate   100  f =       60.309  |proj g|=      0.060883
#> At iterate   101  f =       60.309  |proj g|=      0.047368
#> final  value 60.309243 
#> stopped after 101 iterations
#> 
#> Call:
#> DiceKriging::km(design = data, response = task$truth(), control = pv$control)
#> 
#> Trend  coeff.:
#>                Estimate
#>  (Intercept)    20.6188
#> 
#> Covar. type  : matern5_2 
#> Covar. coeff.:
#>                Estimate
#>    theta(am)     0.0000
#>  theta(carb)     3.3989
#>   theta(cyl)     1.0795
#>  theta(disp)   468.1214
#>  theta(drat)     0.7911
#>  theta(gear)     3.2118
#>    theta(hp)   354.8848
#>  theta(qsec)     0.5468
#>    theta(vs)     2.0000
#>    theta(wt)     6.9265
#> 
#> Variance estimate: 29.821
#> regr.mse 
#> 15.61579