Neural Network Regression Learner

Single Layer Neural Network. Calls nnet::nnet.formula() from package nnet.

Note that modern neural networks with multiple layers are connected via package mlr3torch.

Dictionary

This mlr3::Learner can be instantiated via the dictionary mlr3::mlr_learners or with the associated sugar function mlr3::lrn():

mlr_learners$get("regr.nnet")
lrn("regr.nnet")

Meta Information

Task type: “regr”

Predict Types: “response”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3learners, nnet

Parameters

Id	Type	Default	Levels	Range

Hess	logical	FALSE	TRUE, FALSE	-

MaxNWts	integer	1000		\([1, \infty)\)

Wts	untyped	-		-

abstol	numeric	1e-04		\((-\infty, \infty)\)

censored	logical	FALSE	TRUE, FALSE	-

contrasts	untyped	NULL		-

decay	numeric	0		\((-\infty, \infty)\)

mask	untyped	-		-

maxit	integer	100		\([1, \infty)\)

na.action	untyped	-		-

rang	numeric	0.7		\((-\infty, \infty)\)

reltol	numeric	1e-08		\((-\infty, \infty)\)

size	integer	3		\([0, \infty)\)

skip	logical	FALSE	TRUE, FALSE	-

subset	untyped	-		-

trace	logical	TRUE	TRUE, FALSE	-

formula	untyped	-		-

Initial parameter values

size:

Adjusted default: 3L.

Reason for change: no default in nnet().

Custom mlr3 parameters

formula: if not provided, the formula is set to task$formula().

References

Ripley BD (1996). Pattern Recognition and Neural Networks. Cambridge University Press. doi:10.1017/cbo9780511812651 .

See also

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrNnet

Methods

Public methods

LearnerRegrNnet$new()

LearnerRegrNnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

mlr3::LearnerRegr$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrNnet$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrNnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner and set parameter values
learner = lrn("regr.nnet")
print(learner)
#> 
#> ── <LearnerRegrNnet> (regr.nnet): Single Layer Neural Network ──────────────────
#> • Model: -
#> • Parameters: size=3
#> • Packages: mlr3, mlr3learners, and nnet
#> • Predict Types: [response]
#> • Feature Types: logical, integer, numeric, factor, and ordered
#> • Encapsulation: none (fallback: -)
#> • Properties: weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)
#> # weights:  37
#> initial  value 9370.049674 
#> final  value 735.418095 
#> converged

# Print the model
print(learner$model)
#> a 10-3-1 network with 37 weights
#> inputs: am carb cyl disp drat gear hp qsec vs wt 
#> output(s): mpg 
#> options were - linear output units 

# Importance method
if ("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> regr.mse 
#> 36.19762