Ranger Regression Learner

Random regression forest. Calls ranger::ranger() from package ranger.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.ranger")
lrn("regr.ranger")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “logical”, “integer”, “numeric”, “character”, “factor”, “ordered”

Required Packages: mlr3, mlr3learners, ranger

Parameters

Id	Type	Default	Range	Levels

alpha	numeric	0.5	\((-\infty, \infty)\)	-

always.split.variables	list	-	-	-

holdout	logical	FALSE	-	TRUE, FALSE

importance	character	-	-	none, impurity, impurity_corrected, permutation

keep.inbag	logical	FALSE	-	TRUE, FALSE

max.depth	integer	NULL	\([0, \infty)\)	-

min.node.size	integer	5	\([1, \infty)\)	-

min.prop	numeric	0.1	\((-\infty, \infty)\)	-

minprop	numeric	0.1	\((-\infty, \infty)\)	-

mtry	integer	-	\([1, \infty)\)	-

mtry.ratio	numeric	-	\([0, 1]\)	-

num.random.splits	integer	1	\([1, \infty)\)	-

num.threads	integer	1	\([1, \infty)\)	-

num.trees	integer	500	\([1, \infty)\)	-

oob.error	logical	TRUE	-	TRUE, FALSE

quantreg	logical	FALSE	-	TRUE, FALSE

regularization.factor	list	1	-	-

regularization.usedepth	logical	FALSE	-	TRUE, FALSE

replace	logical	TRUE	-	TRUE, FALSE

respect.unordered.factors	character	ignore	-	ignore, order, partition

sample.fraction	numeric	-	\([0, 1]\)	-

save.memory	logical	FALSE	-	TRUE, FALSE

scale.permutation.importance	logical	FALSE	-	TRUE, FALSE

se.method	character	infjack	-	jack, infjack

seed	integer	NULL	\((-\infty, \infty)\)	-

split.select.weights	list	NULL	-	-

splitrule	character	variance	-	variance, extratrees, maxstat

verbose	logical	TRUE	-	TRUE, FALSE

write.forest	logical	TRUE	-	TRUE, FALSE

Custom mlr3 defaults

num.threads:

Actual default: NULL, triggering auto-detection of the number of CPUs.

Adjusted value: 1.

Reason for change: Conflicting with parallelization via future.

mtry:

This hyperparameter can alternatively be set via our hyperparameter mtry.ratio as mtry = max(ceiling(mtry.ratio * n_features), 1). Note that mtry and mtry.ratio are mutually exclusive.

References

Wright, N. M, Ziegler, Andreas (2017). “ranger: A Fast Implementation of Random Forests for High Dimensional Data in C++ and R.” Journal of Statistical Software, 77(1), 1--17. doi: 10.18637/jss.v077.i01 .

Breiman, Leo (2001). “Random Forests.” Machine Learning, 45(1), 5--32. ISSN 1573-0565, doi: 10.1023/A:1010933404324 .

See also

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrRanger

Methods

Public methods

LearnerRegrRanger$new()

LearnerRegrRanger$importance()

LearnerRegrRanger$oob_error()

LearnerRegrRanger$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrRanger$new()

Method importance()

The importance scores are extracted from the model slot variable.importance. Parameter importance.mode must be set to "impurity", "impurity_corrected", or "permutation"

Usage

LearnerRegrRanger$importance()

Returns

Named numeric().

Method oob_error()

The out-of-bag error, extracted from model slot prediction.error.

Usage

LearnerRegrRanger$oob_error()

Returns

numeric(1).

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrRanger$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("ranger", quietly = TRUE)) {
  learner = mlr3::lrn("regr.ranger")
  print(learner)

  # available parameters:
learner$param_set$ids()
}
#> <LearnerRegrRanger:regr.ranger>
#> * Model: -
#> * Parameters: num.threads=1
#> * Packages: mlr3, mlr3learners, ranger
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, character, factor, ordered
#> * Properties: hotstart_backward, importance, oob_error, weights
#>  [1] "alpha"                        "always.split.variables"      
#>  [3] "holdout"                      "importance"                  
#>  [5] "keep.inbag"                   "max.depth"                   
#>  [7] "min.node.size"                "min.prop"                    
#>  [9] "minprop"                      "mtry"                        
#> [11] "mtry.ratio"                   "num.random.splits"           
#> [13] "num.threads"                  "num.trees"                   
#> [15] "oob.error"                    "quantreg"                    
#> [17] "regularization.factor"        "regularization.usedepth"     
#> [19] "replace"                      "respect.unordered.factors"   
#> [21] "sample.fraction"              "save.memory"                 
#> [23] "scale.permutation.importance" "se.method"                   
#> [25] "seed"                         "split.select.weights"        
#> [27] "splitrule"                    "verbose"                     
#> [29] "write.forest"