GLM with Elastic Net Regularization Classification Learner

Generalized linear models with elastic net regularization. Calls glmnet::cv.glmnet() from package glmnet.

The default for hyperparameter family is set to "binomial" or "multinomial", depending on the number of classes.

Offset

If a Task contains a column with the offset role, it is automatically incorporated during training via the offset argument in glmnet::glmnet(). During prediction, the offset column from the test set is used only if use_pred_offset = TRUE (default), passed via the newoffset argument in glmnet::predict.glmnet(). Otherwise, if the user sets use_pred_offset = FALSE, a zero offset is applied, effectively disabling the offset adjustment during prediction.

Dictionary

This mlr3::Learner can be instantiated via the dictionary mlr3::mlr_learners or with the associated sugar function mlr3::lrn():

mlr_learners$get("classif.cv_glmnet")
lrn("classif.cv_glmnet")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3learners, glmnet

Parameters

Id	Type	Default	Levels	Range

alignment	character	lambda	lambda, fraction	-

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

epsnr	numeric	1e-08		\([0, 1]\)

eps	numeric	1e-06		\([0, 1]\)

exclude	integer	-		\([1, \infty)\)

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

foldid	untyped	NULL		-

gamma	untyped	-		-

grouped	logical	TRUE	TRUE, FALSE	-

intercept	logical	TRUE	TRUE, FALSE	-

keep	logical	FALSE	TRUE, FALSE	-

lambda.min.ratio	numeric	-		\([0, 1]\)

lambda	untyped	-		-

lower.limits	untyped	-		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

nfolds	integer	10		\([3, \infty)\)

nlambda	integer	100		\([1, \infty)\)

use_pred_offset	logical	-	TRUE, FALSE	-

parallel	logical	FALSE	TRUE, FALSE	-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

predict.gamma	numeric	gamma.1se		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	lambda.1se		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	-	Newton, modified.Newton	-

type.measure	character	deviance	deviance, class, auc, mse, mae	-

type.multinomial	character	-	ungrouped, grouped	-

upper.limits	untyped	-		-

Internal Encoding

Starting with mlr3 v0.5.0, the order of class labels is reversed prior to model fitting to comply to the stats::glm() convention that the negative class is provided as the first factor level.

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1–22. doi:10.18637/jss.v033.i01 .

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifCVGlmnet

Methods

Public methods

LearnerClassifCVGlmnet$new()

LearnerClassifCVGlmnet$selected_features()

LearnerClassifCVGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

mlr3::LearnerClassif$predict_newdata_fast()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifCVGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerClassifCVGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifCVGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

# Define the Learner and set parameter values
learner = lrn("classif.cv_glmnet")
print(learner)
#> 
#> ── <LearnerClassifCVGlmnet> (classif.cv_glmnet): GLM with Elastic Net Regulariza
#> • Model: -
#> • Parameters: use_pred_offset=TRUE
#> • Packages: mlr3, mlr3learners, and glmnet
#> • Predict Types: [response] and prob
#> • Feature Types: logical, integer, and numeric
#> • Encapsulation: none (fallback: -)
#> • Properties: multiclass, offset, selected_features, twoclass, and weights
#> • Other settings: use_weights = 'use'

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# Print the model
print(learner$model)
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "binomial") 
#> 
#> Measure: Binomial Deviance 
#> 
#>      Lambda Index Measure      SE Nonzero
#> min 0.03215    23  0.9976 0.07175      17
#> 1se 0.07427    14  1.0581 0.03836      10

# Importance method
if ("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
#> classif.ce 
#>  0.3043478