GLM with Elastic Net Regularization Classification Learner

Generalized linear models with elastic net regularization. Calls glmnet::glmnet() from package glmnet.

Details

Caution: This learner is different to learners calling glmnet::cv.glmnet() in that it does not use the internal optimization of parameter lambda. Instead, lambda needs to be tuned by the user (e.g., via mlr3tuning). When lambda is tuned, the glmnet will be trained for each tuning iteration. While fitting the whole path of lambdas would be more efficient, as is done by default in glmnet::glmnet(), tuning/selecting the parameter at prediction time (using parameter s) is currently not supported in mlr3 (at least not in efficient manner). Tuning the s parameter is, therefore, currently discouraged.

When the data are i.i.d. and efficiency is key, we recommend using the respective auto-tuning counterparts in mlr_learners_classif.cv_glmnet() or mlr_learners_regr.cv_glmnet(). However, in some situations this is not applicable, usually when data are imbalanced or not i.i.d. (longitudinal, time-series) and tuning requires custom resampling strategies (blocked design, stratification).

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.glmnet")
lrn("classif.glmnet")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3learners, glmnet

Parameters

Id	Type	Default	Levels	Range

alpha	numeric	1		\([0, 1]\)

big	numeric	9.9e+35		\((-\infty, \infty)\)

devmax	numeric	0.999		\([0, 1]\)

dfmax	integer	-		\([0, \infty)\)

eps	numeric	1e-06		\([0, 1]\)

epsnr	numeric	1e-08		\([0, 1]\)

exact	logical	FALSE	TRUE, FALSE	-

exclude	integer	-		\([1, \infty)\)

exmx	numeric	250		\((-\infty, \infty)\)

fdev	numeric	1e-05		\([0, 1]\)

gamma	numeric	1		\((-\infty, \infty)\)

intercept	logical	TRUE	TRUE, FALSE	-

lambda	untyped	-		-

lambda.min.ratio	numeric	-		\([0, 1]\)

lower.limits	untyped	-		-

maxit	integer	100000		\([1, \infty)\)

mnlam	integer	5		\([1, \infty)\)

mxit	integer	100		\([1, \infty)\)

mxitnr	integer	25		\([1, \infty)\)

nlambda	integer	100		\([1, \infty)\)

newoffset	untyped	-		-

offset	untyped			-

penalty.factor	untyped	-		-

pmax	integer	-		\([0, \infty)\)

pmin	numeric	1e-09		\([0, 1]\)

prec	numeric	1e-10		\((-\infty, \infty)\)

relax	logical	FALSE	TRUE, FALSE	-

s	numeric	0.01		\([0, \infty)\)

standardize	logical	TRUE	TRUE, FALSE	-

standardize.response	logical	FALSE	TRUE, FALSE	-

thresh	numeric	1e-07		\([0, \infty)\)

trace.it	integer	0		\([0, 1]\)

type.gaussian	character	-	covariance, naive	-

type.logistic	character	-	Newton, modified.Newton	-

type.multinomial	character	-	ungrouped, grouped	-

upper.limits	untyped	-		-

Internal Encoding

Starting with mlr3 v0.5.0, the order of class labels is reversed prior to model fitting to comply to the stats::glm() convention that the negative class is provided as the first factor level.

References

Friedman J, Hastie T, Tibshirani R (2010). “Regularization Paths for Generalized Linear Models via Coordinate Descent.” Journal of Statistical Software, 33(1), 1--22. doi:10.18637/jss.v033.i01 .

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifGlmnet

Methods

Public methods

LearnerClassifGlmnet$new()

LearnerClassifGlmnet$selected_features()

LearnerClassifGlmnet$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifGlmnet$new()

Method selected_features()

Returns the set of selected features as reported by glmnet::predict.glmnet() with type set to "nonzero".

Usage

LearnerClassifGlmnet$selected_features(lambda = NULL)

Arguments

lambda

(numeric(1))
Custom lambda, defaults to the active lambda depending on parameter set.

Returns

(character()) of feature names.

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifGlmnet$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("glmnet", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("classif.glmnet")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerClassifGlmnet:classif.glmnet>: GLM with Elastic Net Regularization
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, glmnet
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric
#> * Properties: multiclass, twoclass, weights
#> 
#> Call:  (if (cv) glmnet::cv.glmnet else glmnet::glmnet)(x = data, y = target,      family = "binomial") 
#> 
#>     Df  %Dev   Lambda
#> 1    0  0.00 0.239600
#> 2    1  2.85 0.218300
#> 3    1  5.26 0.198900
#> 4    1  7.32 0.181300
#> 5    2  9.23 0.165200
#> 6    3 11.17 0.150500
#> 7    5 14.11 0.137100
#> 8    6 17.31 0.124900
#> 9    6 20.50 0.113800
#> 10   6 23.29 0.103700
#> 11   6 25.73 0.094510
#> 12   6 27.90 0.086110
#> 13   6 29.82 0.078460
#> 14   8 31.66 0.071490
#> 15   8 33.43 0.065140
#> 16   8 35.01 0.059350
#> 17  12 36.71 0.054080
#> 18  12 38.75 0.049280
#> 19  12 40.57 0.044900
#> 20  14 42.30 0.040910
#> 21  15 43.97 0.037280
#> 22  18 45.55 0.033960
#> 23  18 47.01 0.030950
#> 24  18 48.32 0.028200
#> 25  18 49.51 0.025690
#> 26  21 51.22 0.023410
#> 27  22 53.03 0.021330
#> 28  22 54.80 0.019440
#> 29  25 57.00 0.017710
#> 30  26 59.16 0.016140
#> 31  27 61.13 0.014700
#> 32  28 62.99 0.013400
#> 33  28 64.69 0.012210
#> 34  30 66.36 0.011120
#> 35  32 67.98 0.010130
#> 36  32 69.58 0.009234
#> 37  32 71.06 0.008413
#> 38  35 72.50 0.007666
#> 39  35 73.91 0.006985
#> 40  37 75.29 0.006364
#> 41  36 76.68 0.005799
#> 42  35 77.98 0.005284
#> 43  36 79.22 0.004814
#> 44  35 80.37 0.004387
#> 45  36 81.51 0.003997
#> 46  37 82.60 0.003642
#> 47  37 83.66 0.003318
#> 48  37 84.68 0.003024
#> 49  38 85.73 0.002755
#> 50  38 86.70 0.002510
#> 51  38 87.65 0.002287
#> 52  39 88.60 0.002084
#> 53  40 89.49 0.001899
#> 54  41 90.35 0.001730
#> 55  39 91.17 0.001577
#> 56  39 91.90 0.001436
#> 57  41 92.58 0.001309
#> 58  41 93.22 0.001193
#> 59  41 93.81 0.001087
#> 60  42 94.35 0.000990
#> 61  42 94.85 0.000902
#> 62  42 95.30 0.000822
#> 63  42 95.71 0.000749
#> 64  42 96.09 0.000682
#> 65  42 96.43 0.000622
#> 66  41 96.75 0.000567
#> 67  42 97.03 0.000516
#> 68  42 97.30 0.000470
#> 69  42 97.53 0.000429
#> 70  42 97.75 0.000390
#> 71  42 97.95 0.000356
#> 72  43 98.13 0.000324
#> 73  42 98.30 0.000295
#> 74  42 98.45 0.000269
#> 75  42 98.58 0.000245
#> 76  42 98.71 0.000224
#> 77  43 98.82 0.000204
#> 78  43 98.92 0.000186
#> 79  43 99.02 0.000169
#> 80  43 99.10 0.000154
#> 81  43 99.18 0.000140
#> 82  43 99.25 0.000128
#> 83  43 99.32 0.000116
#> 84  43 99.38 0.000106
#> 85  44 99.43 0.000097
#> 86  44 99.48 0.000088
#> 87  44 99.53 0.000080
#> 88  44 99.57 0.000073
#> 89  44 99.61 0.000067
#> 90  44 99.64 0.000061
#> 91  44 99.68 0.000055
#> 92  44 99.70 0.000050
#> 93  44 99.73 0.000046
#> 94  44 99.75 0.000042
#> 95  44 99.78 0.000038
#> 96  44 99.79 0.000035
#> 97  44 99.81 0.000032
#> 98  44 99.83 0.000029
#> 99  44 99.84 0.000026
#> 100 44 99.86 0.000024
#> Warning: Multiple lambdas have been fit. Lambda will be set to 0.01 (see parameter 's').
#> classif.ce 
#>  0.3478261