Linear Discriminant Analysis Classification Learner

Linear discriminant analysis. Calls MASS::lda() from package MASS.

Details

Parameters method and prior exist for training and prediction but accept different values for each. Therefore, arguments for the predict stage have been renamed to predict.method and predict.prior, respectively.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.lda")
lrn("classif.lda")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3learners, MASS

Parameters

Id	Type	Default	Levels	Range

dimen	untyped	-		-

method	character	moment	moment, mle, mve, t	-

nu	integer	-		\((-\infty, \infty)\)

predict.method	character	plug-in	plug-in, predictive, debiased	-

predict.prior	untyped	-		-

prior	untyped	-		-

tol	numeric	-		\((-\infty, \infty)\)

References

Venables WN, Ripley BD (2002). Modern Applied Statistics with S, Fourth edition. Springer, New York. ISBN 0-387-95457-0, http://www.stats.ox.ac.uk/pub/MASS4/.

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLDA

Methods

Public methods

LearnerClassifLDA$new()

LearnerClassifLDA$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLDA$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLDA$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("MASS", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("classif.lda")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerClassifLDA:classif.lda>: Linear Discriminant Analysis
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, MASS
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor, ordered
#> * Properties: multiclass, twoclass, weights
#> Call:
#> lda(formula, data = task$data())
#> 
#> Prior probabilities of groups:
#>         M         R 
#> 0.5323741 0.4676259 
#> 
#> Group means:
#>           V1       V10       V11       V12       V13       V14       V15
#> M 0.03692162 0.2625973 0.2998297 0.3067676 0.3233189 0.3295122 0.3307662
#> R 0.02139231 0.1813969 0.1895754 0.1954431 0.2212015 0.2631138 0.3079262
#>         V16       V17       V18       V19         V2       V20       V21
#> M 0.3841041 0.4145527 0.4534014 0.5334797 0.04469189 0.6348392 0.7021554
#> R 0.3830062 0.4165415 0.4362185 0.4518000 0.02949385 0.4874354 0.5348015
#>         V22       V23       V24       V25       V26       V27       V28
#> M 0.6997459 0.6845932 0.6862270 0.6802041 0.7142838 0.7222919 0.7171419
#> R 0.5737000 0.6393938 0.6678846 0.6551508 0.6811538 0.6877323 0.6794908
#>         V29         V3       V30       V31       V32       V33       V34
#> M 0.6426554 0.04807568 0.5742757 0.4759284 0.4355784 0.4013905 0.3689486
#> R 0.6320523 0.03834462 0.5707246 0.5220892 0.4575169 0.4368185 0.4402908
#>         V35       V36       V37       V38       V39         V4       V40
#> M 0.3366243 0.3227959 0.3206851 0.3341743 0.3371865 0.06201216 0.2878257
#> R 0.4671385 0.4758262 0.4174323 0.3366338 0.3064077 0.04533077 0.3147692
#>         V41       V42       V43       V44       V45       V46        V47
#> M 0.2680095 0.2838324 0.2737230 0.2483446 0.2439162 0.1946338 0.13940946
#> R 0.2825385 0.2340785 0.2045323 0.1746523 0.1378000 0.1102892 0.08739692
#>          V48        V49         V5        V50        V51        V52         V53
#> M 0.10911216 0.06438919 0.08495811 0.02325676 0.01927568 0.01496351 0.011059459
#> R 0.06256923 0.03552615 0.06393077 0.01784769 0.01206769 0.01064000 0.009452308
#>           V54         V55         V56         V57         V58         V59
#> M 0.011737838 0.009948649 0.009194595 0.007500000 0.009462162 0.008550000
#> R 0.009596923 0.008078462 0.007556923 0.008530769 0.006678462 0.007426154
#>          V6         V60        V7        V8        V9
#> M 0.1137608 0.006901351 0.1303446 0.1496162 0.2222473
#> R 0.1011415 0.006132308 0.1181385 0.1272646 0.1483631
#> 
#> Coefficients of linear discriminants:
#>               LD1
#> V1  -31.208946309
#> V10   3.227588566
#> V11  -0.697529554
#> V12  -0.712921785
#> V13  -5.627164193
#> V14   2.009158350
#> V15  -1.311123272
#> V16   0.726340667
#> V17   4.813764826
#> V18  -4.386949523
#> V19   0.451209314
#> V2   -2.378935162
#> V20   0.727792038
#> V21   0.828058189
#> V22  -2.871319170
#> V23   0.170849234
#> V24  -0.516687196
#> V25   1.328248940
#> V26  -0.654535071
#> V27  -1.165745429
#> V28   0.319588490
#> V29   2.667547713
#> V3   26.452037981
#> V30  -4.132717729
#> V31   3.602374432
#> V32  -0.662268770
#> V33  -0.962668476
#> V34   2.299740349
#> V35  -3.822149293
#> V36   2.142799674
#> V37   1.057316028
#> V38  -0.009112236
#> V39  -4.515921952
#> V4  -11.259612969
#> V40   5.102002668
#> V41  -0.605481204
#> V42   2.215509836
#> V43  -3.251382752
#> V44   2.639641813
#> V45  -1.894576182
#> V46   2.374506094
#> V47 -10.408016549
#> V48  -3.399704501
#> V49 -25.818345307
#> V5   -0.129574969
#> V50  64.538123539
#> V51  -5.176490971
#> V52  -9.014173810
#> V53 -21.272957643
#> V54 -18.077347820
#> V55  34.594437970
#> V56   8.034818764
#> V57  44.828601775
#> V58 -27.542605144
#> V59 -13.839613787
#> V6   -2.259166766
#> V60  10.197672348
#> V7    3.001599304
#> V8    8.948002733
#> V9   -8.937315083
#> classif.ce 
#>  0.2898551