Linear Discriminant Analysis Classification Learner

Linear discriminant analysis. Calls MASS::lda() from package MASS.

Details

Parameters method and prior exist for training and prediction but accept different values for each. Therefore, arguments for the predict stage have been renamed to predict.method and predict.prior, respectively.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.lda")
lrn("classif.lda")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”, “ordered”

Required Packages: mlr3, mlr3learners, MASS

Parameters

Id	Type	Default	Range	Levels

dimen	list	-	-	-

method	character	moment	-	moment, mle, mve, t

nu	integer	-	\((-\infty, \infty)\)	-

predict.method	character	plug-in	-	plug-in, predictive, debiased

predict.prior	list	-	-	-

prior	list	-	-	-

tol	numeric	-	\((-\infty, \infty)\)	-

References

Venables WN, Ripley BD (2002). Modern Applied Statistics with S, Fourth edition. Springer, New York. ISBN 0-387-95457-0, http://www.stats.ox.ac.uk/pub/MASS4/.

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLDA

Methods

Public methods

LearnerClassifLDA$new()

LearnerClassifLDA$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLDA$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLDA$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("MASS", quietly = TRUE)) {
  learner = mlr3::lrn("classif.lda")
  print(learner)

  # available parameters:
learner$param_set$ids()
}
#> <LearnerClassifLDA:classif.lda>
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, MASS
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, factor, ordered
#> * Properties: multiclass, twoclass, weights
#> [1] "dimen"          "method"         "nu"             "predict.method"
#> [5] "predict.prior"  "prior"          "tol"