Logistic Regression Classification Learner

Classification via logistic regression. Calls stats::glm() with family set to "binomial".

Usage,NULL

Internal Encoding

Starting with mlr3 v0.5.0, the order of class labels is reversed prior to model fitting to comply to the stats::glm() convention that the negative class is provided as the first factor level.

Custom mlr3 defaults

model:

Actual default: TRUE.

Adjusted default: FALSE.

Reason for change: Save some memory.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.log_reg")
lrn("classif.log_reg")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “character”, “factor”, “ordered”

Required Packages: mlr3, mlr3learners, 'stats'

Parameters

Id	Type	Default	Range	Levels

dispersion	list	NULL	-	-

epsilon	numeric	1e-08	\((-\infty, \infty)\)	-

etastart	list	-	-	-

maxit	numeric	25	\((-\infty, \infty)\)	-

model	logical	TRUE	-	TRUE, FALSE

mustart	list	-	-	-

offset	list	-	-	-

singular.ok	logical	TRUE	-	TRUE, FALSE

start	list	NULL	-	-

trace	logical	FALSE	-	TRUE, FALSE

x	logical	FALSE	-	TRUE, FALSE

y	logical	TRUE	-	TRUE, FALSE

Contrasts

To ensure reproducibility, this learner always uses the default contrasts:

contr.treatment() for unordered factors, and

contr.poly() for ordered factors.

Setting the option "contrasts" does not have any effect. Instead, set the respective hyperparameter or use mlr3pipelines to create dummy features.

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifLogReg

Methods

Public methods

LearnerClassifLogReg$new()

LearnerClassifLogReg$loglik()

LearnerClassifLogReg$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifLogReg$new()

Method loglik()

Extract the log-likelihood (e.g., via stats::logLik() from the fitted model.

Usage

LearnerClassifLogReg$loglik()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifLogReg$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("stats", quietly = TRUE)) {
  learner = mlr3::lrn("classif.log_reg")
  print(learner)

  # available parameters:
learner$param_set$ids()
}
#> <LearnerClassifLogReg:classif.log_reg>
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, stats
#> * Predict Type: response
#> * Feature types: logical, integer, numeric, character, factor, ordered
#> * Properties: loglik, twoclass, weights
#>  [1] "dispersion"  "epsilon"     "etastart"    "maxit"       "model"      
#>  [6] "mustart"     "offset"      "singular.ok" "start"       "trace"      
#> [11] "x"           "y"