Naive Bayes Classification Learner

Naive Bayes classification. Calls e1071::naiveBayes() from package e1071.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.naive_bayes")
lrn("classif.naive_bayes")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3learners, e1071

Parameters

Id	Type	Default	Range

eps	numeric	0	\((-\infty, \infty)\)

laplace	numeric	0	\([0, \infty)\)

threshold	numeric	0.001	\((-\infty, \infty)\)

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayes

Methods

Public methods

LearnerClassifNaiveBayes$new()

LearnerClassifNaiveBayes$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("e1071", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("classif.naive_bayes")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerClassifNaiveBayes:classif.naive_bayes>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, e1071
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor
#> * Properties: multiclass, twoclass
#> 
#> Naive Bayes Classifier for Discrete Predictors
#> 
#> Call:
#> naiveBayes.default(x = x, y = y)
#> 
#> A-priori probabilities:
#> y
#>         M         R 
#> 0.5323741 0.4676259 
#> 
#> Conditional probabilities:
#>    V1
#> y         [,1]       [,2]
#>   M 0.03738243 0.02885748
#>   R 0.02263538 0.01593063
#> 
#>    V10
#> y        [,1]      [,2]
#>   M 0.2569824 0.1421978
#>   R 0.1773631 0.1260982
#> 
#>    V11
#> y        [,1]      [,2]
#>   M 0.2913014 0.1266952
#>   R 0.1796446 0.1231584
#> 
#>    V12
#> y        [,1]      [,2]
#>   M 0.3067311 0.1276957
#>   R 0.1952031 0.1457309
#> 
#>    V13
#> y        [,1]      [,2]
#>   M 0.3172230 0.1375364
#>   R 0.2287138 0.1498331
#> 
#>    V14
#> y        [,1]      [,2]
#>   M 0.3171743 0.1572874
#>   R 0.2705615 0.1802852
#> 
#>    V15
#> y        [,1]      [,2]
#>   M 0.3275743 0.1873194
#>   R 0.3021892 0.2277690
#> 
#>    V16
#> y        [,1]      [,2]
#>   M 0.3843432 0.2034275
#>   R 0.3667585 0.2651091
#> 
#>    V17
#> y        [,1]      [,2]
#>   M 0.4217527 0.2390404
#>   R 0.4046862 0.2944651
#> 
#>    V18
#> y        [,1]      [,2]
#>   M 0.4518000 0.2608769
#>   R 0.4354123 0.2635662
#> 
#>    V19
#> y        [,1]      [,2]
#>   M 0.5300581 0.2634300
#>   R 0.4378185 0.2527825
#> 
#>    V2
#> y         [,1]       [,2]
#>   M 0.04758919 0.04043565
#>   R 0.03212923 0.02719114
#> 
#>    V20
#> y        [,1]      [,2]
#>   M 0.6187446 0.2548798
#>   R 0.4712908 0.2526204
#> 
#>    V21
#> y        [,1]      [,2]
#>   M 0.6628446 0.2546494
#>   R 0.5266354 0.2506294
#> 
#>    V22
#> y        [,1]      [,2]
#>   M 0.6534730 0.2370555
#>   R 0.5721185 0.2663592
#> 
#>    V23
#> y        [,1]      [,2]
#>   M 0.6510797 0.2494619
#>   R 0.6158200 0.2578687
#> 
#>    V24
#> y        [,1]      [,2]
#>   M 0.6573986 0.2452991
#>   R 0.6270938 0.2493544
#> 
#>    V25
#> y        [,1]      [,2]
#>   M 0.6454986 0.2477617
#>   R 0.6313985 0.2702582
#> 
#>    V26
#> y        [,1]      [,2]
#>   M 0.6768932 0.2415243
#>   R 0.6671123 0.2532012
#> 
#>    V27
#> y        [,1]      [,2]
#>   M 0.6931973 0.2700855
#>   R 0.6728138 0.2225997
#> 
#>    V28
#> y        [,1]      [,2]
#>   M 0.6961311 0.2671151
#>   R 0.6614815 0.2126706
#> 
#>    V29
#> y        [,1]      [,2]
#>   M 0.6342176 0.2564996
#>   R 0.6258846 0.2376075
#> 
#>    V3
#> y         [,1]       [,2]
#>   M 0.05421351 0.04668135
#>   R 0.03732769 0.03117378
#> 
#>    V30
#> y        [,1]      [,2]
#>   M 0.5888270 0.2251604
#>   R 0.5768523 0.2427821
#> 
#>    V31
#> y        [,1]      [,2]
#>   M 0.4939068 0.2379245
#>   R 0.5408877 0.2100200
#> 
#>    V32
#> y        [,1]      [,2]
#>   M 0.4386892 0.2202986
#>   R 0.4725908 0.2098967
#> 
#>    V33
#> y        [,1]      [,2]
#>   M 0.4034500 0.1878922
#>   R 0.4562215 0.2254117
#> 
#>    V34
#> y        [,1]      [,2]
#>   M 0.3821135 0.2015807
#>   R 0.4673738 0.2488202
#> 
#>    V35
#> y        [,1]      [,2]
#>   M 0.3592068 0.2447263
#>   R 0.4849692 0.2650990
#> 
#>    V36
#> y        [,1]      [,2]
#>   M 0.3247324 0.2507964
#>   R 0.4890846 0.2613533
#> 
#>    V37
#> y        [,1]      [,2]
#>   M 0.3204824 0.2189517
#>   R 0.4421154 0.2480089
#> 
#>    V38
#> y        [,1]      [,2]
#>   M 0.3493905 0.2040099
#>   R 0.3724969 0.2381699
#> 
#>    V39
#> y        [,1]      [,2]
#>   M 0.3543703 0.1876785
#>   R 0.3323000 0.2313172
#> 
#>    V4
#> y         [,1]       [,2]
#>   M 0.07067703 0.06067103
#>   R 0.04346615 0.03440655
#> 
#>    V40
#> y        [,1]      [,2]
#>   M 0.3214284 0.1595127
#>   R 0.3442569 0.2154210
#> 
#>    V41
#> y        [,1]      [,2]
#>   M 0.3119486 0.1689144
#>   R 0.3175446 0.1966858
#> 
#>    V42
#> y        [,1]      [,2]
#>   M 0.3239662 0.1714900
#>   R 0.2719631 0.1886174
#> 
#>    V43
#> y        [,1]      [,2]
#>   M 0.2942797 0.1453929
#>   R 0.2303692 0.1403766
#> 
#>    V44
#> y        [,1]      [,2]
#>   M 0.2574716 0.1538779
#>   R 0.1854185 0.1117171
#> 
#>    V45
#> y        [,1]      [,2]
#>   M 0.2626419 0.1875113
#>   R 0.1469277 0.1026594
#> 
#>    V46
#> y        [,1]      [,2]
#>   M 0.2126770 0.1686159
#>   R 0.1167754 0.1025137
#> 
#>    V47
#> y         [,1]       [,2]
#>   M 0.15184865 0.10661670
#>   R 0.09613385 0.07404074
#> 
#>    V48
#> y         [,1]       [,2]
#>   M 0.11398784 0.07197758
#>   R 0.07034308 0.05185432
#> 
#>    V49
#> y         [,1]       [,2]
#>   M 0.06581351 0.03837002
#>   R 0.03873538 0.03208476
#> 
#>    V5
#> y         [,1]       [,2]
#>   M 0.09092297 0.06627116
#>   R 0.06487385 0.04860147
#> 
#>    V50
#> y         [,1]       [,2]
#>   M 0.02397297 0.01604809
#>   R 0.01786769 0.01196648
#> 
#>    V51
#> y         [,1]        [,2]
#>   M 0.02038514 0.015155248
#>   R 0.01194000 0.008965695
#> 
#>    V52
#> y         [,1]        [,2]
#>   M 0.01738649 0.011888746
#>   R 0.01144923 0.007810252
#> 
#>    V53
#> y         [,1]        [,2]
#>   M 0.01218514 0.007916892
#>   R 0.01000615 0.006421421
#> 
#>    V54
#> y         [,1]        [,2]
#>   M 0.01230676 0.008869825
#>   R 0.00936000 0.005336630
#> 
#>    V55
#> y          [,1]        [,2]
#>   M 0.010248649 0.008537335
#>   R 0.008550769 0.005296730
#> 
#>    V56
#> y          [,1]        [,2]
#>   M 0.009647297 0.007134562
#>   R 0.007249231 0.005235579
#> 
#>    V57
#> y          [,1]        [,2]
#>   M 0.007405405 0.005952020
#>   R 0.007924615 0.005958136
#> 
#>    V58
#> y          [,1]        [,2]
#>   M 0.009182432 0.008054937
#>   R 0.007153846 0.005148940
#> 
#>    V59
#> y          [,1]        [,2]
#>   M 0.008813514 0.007237163
#>   R 0.007069231 0.005125912
#> 
#>    V6
#> y        [,1]       [,2]
#>   M 0.1113689 0.05344842
#>   R 0.1019923 0.07220942
#> 
#>    V60
#> y          [,1]        [,2]
#>   M 0.006954054 0.005010706
#>   R 0.006196923 0.003680564
#> 
#>    V7
#> y        [,1]       [,2]
#>   M 0.1225135 0.05431124
#>   R 0.1163031 0.07334022
#> 
#>    V8
#> y        [,1]       [,2]
#>   M 0.1492595 0.08564225
#>   R 0.1286046 0.09029862
#> 
#>    V9
#> y        [,1]      [,2]
#>   M 0.2203905 0.1252937
#>   R 0.1450785 0.1145228
#> 
#> classif.ce 
#>  0.3188406