Naive Bayes Classification Learner

Naive Bayes classification. Calls e1071::naiveBayes() from package e1071.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.naive_bayes")
lrn("classif.naive_bayes")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3learners, e1071

Parameters

Id	Type	Default	Range

eps	numeric	0	\((-\infty, \infty)\)

laplace	numeric	0	\([0, \infty)\)

threshold	numeric	0.001	\((-\infty, \infty)\)

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayes

Methods

Public methods

LearnerClassifNaiveBayes$new()

LearnerClassifNaiveBayes$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("e1071", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("classif.naive_bayes")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerClassifNaiveBayes:classif.naive_bayes>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, e1071
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor
#> * Properties: multiclass, twoclass
#> 
#> Naive Bayes Classifier for Discrete Predictors
#> 
#> Call:
#> naiveBayes.default(x = x, y = y)
#> 
#> A-priori probabilities:
#> y
#>         M         R 
#> 0.5323741 0.4676259 
#> 
#> Conditional probabilities:
#>    V1
#> y         [,1]       [,2]
#>   M 0.03848378 0.03088998
#>   R 0.02186923 0.01523360
#> 
#>    V10
#> y        [,1]      [,2]
#>   M 0.2493108 0.1353904
#>   R 0.1605892 0.1123774
#> 
#>    V11
#> y        [,1]     [,2]
#>   M 0.2866797 0.114291
#>   R 0.1700169 0.104354
#> 
#>    V12
#> y        [,1]      [,2]
#>   M 0.2991514 0.1230780
#>   R 0.1839446 0.1273444
#> 
#>    V13
#> y        [,1]      [,2]
#>   M 0.3115095 0.1354647
#>   R 0.2316185 0.1337897
#> 
#>    V14
#> y        [,1]      [,2]
#>   M 0.3226851 0.1490304
#>   R 0.2911923 0.1612310
#> 
#>    V15
#> y        [,1]      [,2]
#>   M 0.3415351 0.1915232
#>   R 0.3329785 0.2213029
#> 
#>    V16
#> y        [,1]      [,2]
#>   M 0.3987000 0.2112770
#>   R 0.3881231 0.2548749
#> 
#>    V17
#> y        [,1]      [,2]
#>   M 0.4261689 0.2487771
#>   R 0.4202769 0.2783530
#> 
#>    V18
#> y        [,1]      [,2]
#>   M 0.4661122 0.2666675
#>   R 0.4476862 0.2515749
#> 
#>    V19
#> y        [,1]      [,2]
#>   M 0.5407054 0.2622035
#>   R 0.4824677 0.2301691
#> 
#>    V2
#> y         [,1]       [,2]
#>   M 0.04955135 0.04021758
#>   R 0.03023846 0.02612593
#> 
#>    V20
#> y        [,1]      [,2]
#>   M 0.6269865 0.2638513
#>   R 0.5241923 0.2401031
#> 
#>    V21
#> y        [,1]      [,2]
#>   M 0.6576905 0.2638221
#>   R 0.5594769 0.2333523
#> 
#>    V22
#> y        [,1]      [,2]
#>   M 0.6355054 0.2557065
#>   R 0.5937538 0.2388994
#> 
#>    V23
#> y        [,1]      [,2]
#>   M 0.6354676 0.2753216
#>   R 0.6399985 0.2360244
#> 
#>    V24
#> y        [,1]      [,2]
#>   M 0.6464338 0.2573830
#>   R 0.6880446 0.2315611
#> 
#>    V25
#> y        [,1]      [,2]
#>   M 0.6495919 0.2421602
#>   R 0.6974477 0.2294887
#> 
#>    V26
#> y        [,1]      [,2]
#>   M 0.6771378 0.2316452
#>   R 0.7136600 0.2160085
#> 
#>    V27
#> y        [,1]      [,2]
#>   M 0.6783919 0.2709970
#>   R 0.7151892 0.2217375
#> 
#>    V28
#> y        [,1]      [,2]
#>   M 0.6853770 0.2585274
#>   R 0.6965262 0.2060247
#> 
#>    V29
#> y        [,1]      [,2]
#>   M 0.6295865 0.2587338
#>   R 0.6389938 0.2312083
#> 
#>    V3
#> y         [,1]       [,2]
#>   M 0.05516081 0.04858341
#>   R 0.03542769 0.02855568
#> 
#>    V30
#> y        [,1]      [,2]
#>   M 0.5817338 0.2206823
#>   R 0.5828354 0.2424306
#> 
#>    V31
#> y        [,1]      [,2]
#>   M 0.4807311 0.2262183
#>   R 0.5415200 0.1989757
#> 
#>    V32
#> y        [,1]      [,2]
#>   M 0.4321230 0.2108252
#>   R 0.4721246 0.1978266
#> 
#>    V33
#> y        [,1]      [,2]
#>   M 0.4127608 0.2026001
#>   R 0.4371892 0.2076500
#> 
#>    V34
#> y        [,1]      [,2]
#>   M 0.3862122 0.2119769
#>   R 0.4345554 0.2439954
#> 
#>    V35
#> y        [,1]      [,2]
#>   M 0.3538311 0.2505551
#>   R 0.4605523 0.2567949
#> 
#>    V36
#> y        [,1]      [,2]
#>   M 0.3392135 0.2538254
#>   R 0.4698062 0.2619383
#> 
#>    V37
#> y        [,1]      [,2]
#>   M 0.3383946 0.2357143
#>   R 0.4322415 0.2327138
#> 
#>    V38
#> y        [,1]      [,2]
#>   M 0.3590203 0.2216467
#>   R 0.3475954 0.2068104
#> 
#>    V39
#> y        [,1]      [,2]
#>   M 0.3613365 0.1987988
#>   R 0.3018769 0.2041795
#> 
#>    V4
#> y         [,1]       [,2]
#>   M 0.06880946 0.06189804
#>   R 0.04009077 0.02552734
#> 
#>    V40
#> y        [,1]      [,2]
#>   M 0.3247000 0.1708245
#>   R 0.3030292 0.1941745
#> 
#>    V41
#> y        [,1]      [,2]
#>   M 0.3098419 0.1622440
#>   R 0.2668538 0.1674439
#> 
#>    V42
#> y        [,1]      [,2]
#>   M 0.3163770 0.1643323
#>   R 0.2408569 0.1537292
#> 
#>    V43
#> y        [,1]      [,2]
#>   M 0.2841189 0.1425141
#>   R 0.2039815 0.1189329
#> 
#>    V44
#> y        [,1]      [,2]
#>   M 0.2466081 0.1434075
#>   R 0.1740077 0.1121191
#> 
#>    V45
#> y        [,1]      [,2]
#>   M 0.2538432 0.1792982
#>   R 0.1443831 0.0995415
#> 
#>    V46
#> y        [,1]       [,2]
#>   M 0.2061514 0.15626709
#>   R 0.1193985 0.09070674
#> 
#>    V47
#> y        [,1]       [,2]
#>   M 0.1509311 0.09782744
#>   R 0.0943600 0.06332331
#> 
#>    V48
#> y         [,1]       [,2]
#>   M 0.11496081 0.07079065
#>   R 0.06721538 0.04574348
#> 
#>    V49
#> y         [,1]       [,2]
#>   M 0.06660000 0.03711234
#>   R 0.03741538 0.02926616
#> 
#>    V5
#> y         [,1]       [,2]
#>   M 0.08928243 0.06654430
#>   R 0.05512769 0.04001911
#> 
#>    V50
#> y         [,1]       [,2]
#>   M 0.02316757 0.01456536
#>   R 0.01793385 0.01198026
#> 
#>    V51
#> y         [,1]       [,2]
#>   M 0.01891486 0.01344394
#>   R 0.01262615 0.00880020
#> 
#>    V52
#> y         [,1]        [,2]
#>   M 0.01637432 0.011290571
#>   R 0.01112000 0.007778339
#> 
#>    V53
#> y          [,1]        [,2]
#>   M 0.012025676 0.008094342
#>   R 0.009313846 0.005698873
#> 
#>    V54
#> y          [,1]        [,2]
#>   M 0.012727027 0.008801125
#>   R 0.009984615 0.005563335
#> 
#>    V55
#> y          [,1]        [,2]
#>   M 0.010433784 0.008443158
#>   R 0.008764615 0.005275961
#> 
#>    V56
#> y          [,1]        [,2]
#>   M 0.009817568 0.006956683
#>   R 0.007575385 0.004855282
#> 
#>    V57
#> y          [,1]        [,2]
#>   M 0.008960811 0.006540303
#>   R 0.008113846 0.005940351
#> 
#>    V58
#> y          [,1]       [,2]
#>   M 0.009763514 0.00778028
#>   R 0.006795385 0.00491501
#> 
#>    V59
#> y          [,1]        [,2]
#>   M 0.009293243 0.007013692
#>   R 0.007367692 0.005408966
#> 
#>    V6
#> y         [,1]       [,2]
#>   M 0.11124459 0.05096783
#>   R 0.08984154 0.06199545
#> 
#>    V60
#> y          [,1]        [,2]
#>   M 0.007495946 0.006741884
#>   R 0.005612308 0.003337828
#> 
#>    V7
#> y        [,1]       [,2]
#>   M 0.1260162 0.06099407
#>   R 0.1113800 0.06085396
#> 
#>    V8
#> y        [,1]       [,2]
#>   M 0.1530419 0.09243021
#>   R 0.1154600 0.07827780
#> 
#>    V9
#> y        [,1]       [,2]
#>   M 0.2177027 0.13275637
#>   R 0.1356354 0.09621329
#> 
#> classif.ce 
#>  0.4057971