Naive Bayes Classification Learner

Naive Bayes classification. Calls e1071::naiveBayes() from package e1071.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.naive_bayes")
lrn("classif.naive_bayes")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3learners, e1071

Parameters

Id	Type	Default	Range

eps	numeric	0	\((-\infty, \infty)\)

laplace	numeric	0	\([0, \infty)\)

threshold	numeric	0.001	\((-\infty, \infty)\)

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayes

Methods

Public methods

LearnerClassifNaiveBayes$new()

LearnerClassifNaiveBayes$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("e1071", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("classif.naive_bayes")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerClassifNaiveBayes:classif.naive_bayes>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, e1071
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor
#> * Properties: multiclass, twoclass
#> 
#> Naive Bayes Classifier for Discrete Predictors
#> 
#> Call:
#> naiveBayes.default(x = x, y = y)
#> 
#> A-priori probabilities:
#> y
#>         M         R 
#> 0.5323741 0.4676259 
#> 
#> Conditional probabilities:
#>    V1
#> y         [,1]       [,2]
#>   M 0.03591486 0.02734027
#>   R 0.02149538 0.01496136
#> 
#>    V10
#> y        [,1]      [,2]
#>   M 0.2418176 0.1303933
#>   R 0.1478585 0.1056595
#> 
#>    V11
#> y        [,1]      [,2]
#>   M 0.2790689 0.1115179
#>   R 0.1683169 0.1140526
#> 
#>    V12
#> y        [,1]      [,2]
#>   M 0.2895365 0.1211542
#>   R 0.1913185 0.1381628
#> 
#>    V13
#> y        [,1]      [,2]
#>   M 0.3086838 0.1306253
#>   R 0.2165138 0.1329480
#> 
#>    V14
#> y        [,1]      [,2]
#>   M 0.3252473 0.1765044
#>   R 0.2708262 0.1547711
#> 
#>    V15
#> y        [,1]      [,2]
#>   M 0.3375189 0.2131581
#>   R 0.3101308 0.2085348
#> 
#>    V16
#> y        [,1]      [,2]
#>   M 0.3919446 0.2287934
#>   R 0.3808554 0.2500391
#> 
#>    V17
#> y        [,1]      [,2]
#>   M 0.4327676 0.2503989
#>   R 0.4267385 0.2918498
#> 
#>    V18
#> y        [,1]      [,2]
#>   M 0.4857351 0.2596174
#>   R 0.4612969 0.2702344
#> 
#>    V19
#> y        [,1]      [,2]
#>   M 0.5641378 0.2569564
#>   R 0.4717000 0.2644859
#> 
#>    V2
#> y         [,1]       [,2]
#>   M 0.04374324 0.03272092
#>   R 0.02734462 0.01931733
#> 
#>    V20
#> y        [,1]      [,2]
#>   M 0.6447473 0.2472669
#>   R 0.5056369 0.2575898
#> 
#>    V21
#> y        [,1]      [,2]
#>   M 0.6880230 0.2357829
#>   R 0.5493923 0.2426005
#> 
#>    V22
#> y        [,1]      [,2]
#>   M 0.6733000 0.2355493
#>   R 0.5720308 0.2611079
#> 
#>    V23
#> y        [,1]      [,2]
#>   M 0.6713041 0.2713039
#>   R 0.6204585 0.2428166
#> 
#>    V24
#> y        [,1]      [,2]
#>   M 0.6826230 0.2707175
#>   R 0.6760831 0.2234221
#> 
#>    V25
#> y        [,1]      [,2]
#>   M 0.6811095 0.2595843
#>   R 0.6944923 0.2337716
#> 
#>    V26
#> y        [,1]      [,2]
#>   M 0.7153000 0.2351892
#>   R 0.6950354 0.2268579
#> 
#>    V27
#> y        [,1]      [,2]
#>   M 0.7158162 0.2642203
#>   R 0.6768062 0.1974952
#> 
#>    V28
#> y        [,1]      [,2]
#>   M 0.7087716 0.2545652
#>   R 0.6713554 0.2076435
#> 
#>    V29
#> y        [,1]      [,2]
#>   M 0.6427027 0.2423331
#>   R 0.6362015 0.2414014
#> 
#>    V3
#> y         [,1]       [,2]
#>   M 0.04835405 0.03769865
#>   R 0.03424615 0.02710720
#> 
#>    V30
#> y        [,1]      [,2]
#>   M 0.5722568 0.1945568
#>   R 0.5741123 0.2404660
#> 
#>    V31
#> y        [,1]      [,2]
#>   M 0.4571851 0.2016666
#>   R 0.5246462 0.2019101
#> 
#>    V32
#> y        [,1]      [,2]
#>   M 0.4014784 0.2012820
#>   R 0.4588369 0.2179686
#> 
#>    V33
#> y        [,1]      [,2]
#>   M 0.3862595 0.1980226
#>   R 0.4417708 0.2248287
#> 
#>    V34
#> y        [,1]      [,2]
#>   M 0.3584041 0.2154734
#>   R 0.4283831 0.2507591
#> 
#>    V35
#> y        [,1]      [,2]
#>   M 0.3405473 0.2479433
#>   R 0.4341046 0.2547006
#> 
#>    V36
#> y        [,1]      [,2]
#>   M 0.3002311 0.2660621
#>   R 0.4473508 0.2646278
#> 
#>    V37
#> y        [,1]      [,2]
#>   M 0.2938932 0.2409783
#>   R 0.4155985 0.2478819
#> 
#>    V38
#> y        [,1]      [,2]
#>   M 0.3244230 0.2214946
#>   R 0.3230508 0.2157671
#> 
#>    V39
#> y        [,1]      [,2]
#>   M 0.3324000 0.1938670
#>   R 0.2928015 0.1982099
#> 
#>    V4
#> y         [,1]       [,2]
#>   M 0.06198243 0.04500549
#>   R 0.04052308 0.02818776
#> 
#>    V40
#> y        [,1]      [,2]
#>   M 0.2991459 0.1689375
#>   R 0.3028831 0.1874117
#> 
#>    V41
#> y        [,1]      [,2]
#>   M 0.2778270 0.1589123
#>   R 0.2784631 0.1565538
#> 
#>    V42
#> y        [,1]      [,2]
#>   M 0.2762919 0.1604282
#>   R 0.2399169 0.1483444
#> 
#>    V43
#> y        [,1]      [,2]
#>   M 0.2547108 0.1301014
#>   R 0.1966985 0.1120551
#> 
#>    V44
#> y        [,1]      [,2]
#>   M 0.2275919 0.1252311
#>   R 0.1647523 0.0891142
#> 
#>    V45
#> y        [,1]       [,2]
#>   M 0.2196622 0.15235685
#>   R 0.1377123 0.08393562
#> 
#>    V46
#> y        [,1]       [,2]
#>   M 0.1802959 0.13306774
#>   R 0.1121985 0.07729501
#> 
#>    V47
#> y         [,1]       [,2]
#>   M 0.13344459 0.08436049
#>   R 0.08881231 0.05447717
#> 
#>    V48
#> y         [,1]       [,2]
#>   M 0.10425811 0.06042312
#>   R 0.06591538 0.04240666
#> 
#>    V49
#> y         [,1]       [,2]
#>   M 0.06187838 0.03230902
#>   R 0.03656308 0.02715671
#> 
#>    V5
#> y         [,1]       [,2]
#>   M 0.08652973 0.05611885
#>   R 0.05145692 0.03996724
#> 
#>    V50
#> y         [,1]       [,2]
#>   M 0.02210676 0.01417650
#>   R 0.01730462 0.01144091
#> 
#>    V51
#> y         [,1]        [,2]
#>   M 0.01913784 0.013155765
#>   R 0.01259692 0.007761141
#> 
#>    V52
#> y         [,1]        [,2]
#>   M 0.01521757 0.009720031
#>   R 0.01044462 0.007161595
#> 
#>    V53
#> y          [,1]        [,2]
#>   M 0.011831081 0.007842926
#>   R 0.009938462 0.006420638
#> 
#>    V54
#> y          [,1]        [,2]
#>   M 0.011672973 0.009011230
#>   R 0.009563077 0.005508816
#> 
#>    V55
#> y          [,1]        [,2]
#>   M 0.010133784 0.009404054
#>   R 0.008412308 0.004917142
#> 
#>    V56
#> y          [,1]        [,2]
#>   M 0.009204054 0.007195898
#>   R 0.007713846 0.004799540
#> 
#>    V57
#> y          [,1]        [,2]
#>   M 0.008029730 0.006334245
#>   R 0.008056923 0.006140337
#> 
#>    V58
#> y          [,1]        [,2]
#>   M 0.008718919 0.007680070
#>   R 0.006187692 0.004596755
#> 
#>    V59
#> y          [,1]        [,2]
#>   M 0.009131081 0.007552899
#>   R 0.006949231 0.005320865
#> 
#>    V6
#> y         [,1]       [,2]
#>   M 0.11276351 0.04822682
#>   R 0.08470308 0.05292557
#> 
#>    V60
#> y          [,1]        [,2]
#>   M 0.007191892 0.006286662
#>   R 0.005767692 0.003379955
#> 
#>    V7
#> y        [,1]       [,2]
#>   M 0.1299500 0.05846975
#>   R 0.1041462 0.05245318
#> 
#>    V8
#> y        [,1]       [,2]
#>   M 0.1534635 0.08879027
#>   R 0.1105846 0.07371080
#> 
#>    V9
#> y        [,1]       [,2]
#>   M 0.2062027 0.12103643
#>   R 0.1281000 0.09051243
#> 
#> classif.ce 
#>  0.3043478