Naive Bayes Classification Learner

Naive Bayes classification. Calls e1071::naiveBayes() from package e1071.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.naive_bayes")
lrn("classif.naive_bayes")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3learners, e1071

Parameters

Id	Type	Default	Range

eps	numeric	0	\((-\infty, \infty)\)

laplace	numeric	0	\([0, \infty)\)

threshold	numeric	0.001	\((-\infty, \infty)\)

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayes

Methods

Public methods

LearnerClassifNaiveBayes$new()

LearnerClassifNaiveBayes$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("e1071", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("classif.naive_bayes")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerClassifNaiveBayes:classif.naive_bayes>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, e1071
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor
#> * Properties: multiclass, twoclass
#> 
#> Naive Bayes Classifier for Discrete Predictors
#> 
#> Call:
#> naiveBayes.default(x = x, y = y)
#> 
#> A-priori probabilities:
#> y
#>         M         R 
#> 0.5323741 0.4676259 
#> 
#> Conditional probabilities:
#>    V1
#> y         [,1]      [,2]
#>   M 0.03455000 0.0272087
#>   R 0.02084769 0.0122850
#> 
#>    V10
#> y        [,1]      [,2]
#>   M 0.2667000 0.1455740
#>   R 0.1518877 0.1065805
#> 
#>    V11
#> y        [,1]      [,2]
#>   M 0.3088230 0.1302860
#>   R 0.1719708 0.1137913
#> 
#>    V12
#> y        [,1]      [,2]
#>   M 0.3144203 0.1212263
#>   R 0.1883015 0.1412554
#> 
#>    V13
#> y        [,1]      [,2]
#>   M 0.3174270 0.1371331
#>   R 0.2265554 0.1480860
#> 
#>    V14
#> y        [,1]      [,2]
#>   M 0.3219270 0.1731036
#>   R 0.2548138 0.1743047
#> 
#>    V15
#> y        [,1]      [,2]
#>   M 0.3419338 0.2024634
#>   R 0.2845369 0.2277962
#> 
#>    V16
#> y        [,1]      [,2]
#>   M 0.3836500 0.2122893
#>   R 0.3555431 0.2637440
#> 
#>    V17
#> y        [,1]      [,2]
#>   M 0.4124541 0.2288397
#>   R 0.4098938 0.2914357
#> 
#>    V18
#> y        [,1]      [,2]
#>   M 0.4466959 0.2426676
#>   R 0.4537585 0.2737203
#> 
#>    V19
#> y        [,1]      [,2]
#>   M 0.5235932 0.2360846
#>   R 0.4815092 0.2649247
#> 
#>    V2
#> y         [,1]       [,2]
#>   M 0.04626351 0.03934988
#>   R 0.02978154 0.02691680
#> 
#>    V20
#> y        [,1]      [,2]
#>   M 0.6015500 0.2449711
#>   R 0.5162369 0.2549753
#> 
#>    V21
#> y        [,1]      [,2]
#>   M 0.6577770 0.2463835
#>   R 0.5498892 0.2440793
#> 
#>    V22
#> y        [,1]      [,2]
#>   M 0.6648730 0.2436120
#>   R 0.5915969 0.2403192
#> 
#>    V23
#> y        [,1]      [,2]
#>   M 0.6657473 0.2555751
#>   R 0.6253431 0.2408816
#> 
#>    V24
#> y        [,1]      [,2]
#>   M 0.6807581 0.2515112
#>   R 0.6410985 0.2201747
#> 
#>    V25
#> y        [,1]      [,2]
#>   M 0.6801622 0.2479608
#>   R 0.6536646 0.2464117
#> 
#>    V26
#> y        [,1]      [,2]
#>   M 0.7203811 0.2383744
#>   R 0.6963000 0.2350909
#> 
#>    V27
#> y        [,1]      [,2]
#>   M 0.7320554 0.2643459
#>   R 0.6948585 0.2240412
#> 
#>    V28
#> y        [,1]      [,2]
#>   M 0.7243784 0.2541199
#>   R 0.6741738 0.2279161
#> 
#>    V29
#> y        [,1]      [,2]
#>   M 0.6638770 0.2359373
#>   R 0.6404154 0.2272112
#> 
#>    V3
#> y         [,1]       [,2]
#>   M 0.04860135 0.04474867
#>   R 0.03485692 0.02946000
#> 
#>    V30
#> y        [,1]      [,2]
#>   M 0.5860851 0.2115876
#>   R 0.5965569 0.2310655
#> 
#>    V31
#> y        [,1]      [,2]
#>   M 0.4856081 0.2110963
#>   R 0.5432738 0.1955487
#> 
#>    V32
#> y        [,1]      [,2]
#>   M 0.4230162 0.2024887
#>   R 0.4709046 0.1964321
#> 
#>    V33
#> y        [,1]      [,2]
#>   M 0.3946108 0.1916978
#>   R 0.4543538 0.2110210
#> 
#>    V34
#> y        [,1]      [,2]
#>   M 0.3637135 0.2060198
#>   R 0.4589708 0.2488840
#> 
#>    V35
#> y        [,1]      [,2]
#>   M 0.3308243 0.2549905
#>   R 0.4662000 0.2649177
#> 
#>    V36
#> y        [,1]      [,2]
#>   M 0.3228081 0.2529288
#>   R 0.4817231 0.2485258
#> 
#>    V37
#> y        [,1]      [,2]
#>   M 0.3299365 0.2311053
#>   R 0.4406492 0.2214669
#> 
#>    V38
#> y        [,1]      [,2]
#>   M 0.3418108 0.1993965
#>   R 0.3597169 0.2222514
#> 
#>    V39
#> y        [,1]      [,2]
#>   M 0.3399946 0.1833116
#>   R 0.3188554 0.2249779
#> 
#>    V4
#> y         [,1]       [,2]
#>   M 0.06312297 0.05724651
#>   R 0.03948923 0.02892586
#> 
#>    V40
#> y        [,1]      [,2]
#>   M 0.3013486 0.1577326
#>   R 0.3276169 0.2077547
#> 
#>    V41
#> y        [,1]      [,2]
#>   M 0.2798000 0.1523958
#>   R 0.2868523 0.1872606
#> 
#>    V42
#> y        [,1]      [,2]
#>   M 0.2959027 0.1673801
#>   R 0.2533154 0.1754383
#> 
#>    V43
#> y        [,1]      [,2]
#>   M 0.2881662 0.1438227
#>   R 0.2150523 0.1400596
#> 
#>    V44
#> y        [,1]      [,2]
#>   M 0.2620338 0.1443584
#>   R 0.1790554 0.1179871
#> 
#>    V45
#> y        [,1]      [,2]
#>   M 0.2535635 0.1712921
#>   R 0.1526908 0.1075297
#> 
#>    V46
#> y        [,1]      [,2]
#>   M 0.1958500 0.1484127
#>   R 0.1247123 0.1047267
#> 
#>    V47
#> y         [,1]       [,2]
#>   M 0.14333514 0.09444695
#>   R 0.09978615 0.07289028
#> 
#>    V48
#> y         [,1]       [,2]
#>   M 0.10992973 0.06742258
#>   R 0.07377385 0.04847435
#> 
#>    V49
#> y         [,1]       [,2]
#>   M 0.06580135 0.03695285
#>   R 0.03664154 0.02980866
#> 
#>    V5
#> y         [,1]       [,2]
#>   M 0.08070946 0.05710869
#>   R 0.06255846 0.04387406
#> 
#>    V50
#> y         [,1]       [,2]
#>   M 0.02308919 0.01291048
#>   R 0.01738769 0.01252741
#> 
#>    V51
#> y         [,1]        [,2]
#>   M 0.01956892 0.014739007
#>   R 0.01253692 0.009522155
#> 
#>    V52
#> y         [,1]        [,2]
#>   M 0.01607432 0.011569647
#>   R 0.01054923 0.007960903
#> 
#>    V53
#> y          [,1]        [,2]
#>   M 0.011862162 0.007472446
#>   R 0.009215385 0.005939019
#> 
#>    V54
#> y          [,1]        [,2]
#>   M 0.012137838 0.007650368
#>   R 0.009581538 0.005237631
#> 
#>    V55
#> y          [,1]        [,2]
#>   M 0.009668919 0.007216342
#>   R 0.008764615 0.005456230
#> 
#>    V56
#> y          [,1]        [,2]
#>   M 0.008013514 0.004897534
#>   R 0.007735385 0.004880716
#> 
#>    V57
#> y          [,1]        [,2]
#>   M 0.007412162 0.005099126
#>   R 0.007520000 0.005639625
#> 
#>    V58
#> y          [,1]        [,2]
#>   M 0.008885135 0.007235485
#>   R 0.006487692 0.005074585
#> 
#>    V59
#> y          [,1]       [,2]
#>   M 0.007906757 0.00589391
#>   R 0.006658462 0.00400706
#> 
#>    V6
#> y         [,1]       [,2]
#>   M 0.10888649 0.05364562
#>   R 0.09465077 0.06144011
#> 
#>    V60
#> y          [,1]        [,2]
#>   M 0.006629730 0.005034613
#>   R 0.006146154 0.003698990
#> 
#>    V7
#> y        [,1]       [,2]
#>   M 0.1329365 0.06198786
#>   R 0.1099169 0.05942239
#> 
#>    V8
#> y        [,1]       [,2]
#>   M 0.1537892 0.08598019
#>   R 0.1117646 0.07735980
#> 
#>    V9
#> y        [,1]       [,2]
#>   M 0.2235041 0.12284741
#>   R 0.1277092 0.09712549
#> 
#> classif.ce 
#>  0.2898551