Naive Bayes Classification Learner

Naive Bayes classification. Calls e1071::naiveBayes() from package e1071.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.naive_bayes")
lrn("classif.naive_bayes")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3learners, e1071

Parameters

Id	Type	Default	Range

eps	numeric	0	\((-\infty, \infty)\)

laplace	numeric	0	\([0, \infty)\)

threshold	numeric	0.001	\((-\infty, \infty)\)

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayes

Methods

Public methods

LearnerClassifNaiveBayes$new()

LearnerClassifNaiveBayes$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("e1071", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("classif.naive_bayes")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerClassifNaiveBayes:classif.naive_bayes>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, e1071
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor
#> * Properties: multiclass, twoclass
#> 
#> Naive Bayes Classifier for Discrete Predictors
#> 
#> Call:
#> naiveBayes.default(x = x, y = y)
#> 
#> A-priori probabilities:
#> y
#>         M         R 
#> 0.5323741 0.4676259 
#> 
#> Conditional probabilities:
#>    V1
#> y         [,1]       [,2]
#>   M 0.03210811 0.02535078
#>   R 0.02103538 0.01161506
#> 
#>    V10
#> y        [,1]      [,2]
#>   M 0.2445122 0.1278422
#>   R 0.1485215 0.1079859
#> 
#>    V11
#> y        [,1]       [,2]
#>   M 0.2871689 0.11216692
#>   R 0.1602538 0.09993849
#> 
#>    V12
#> y        [,1]      [,2]
#>   M 0.3028216 0.1179572
#>   R 0.1674538 0.1150310
#> 
#>    V13
#> y        [,1]      [,2]
#>   M 0.3168743 0.1301517
#>   R 0.2014031 0.1223038
#> 
#>    V14
#> y        [,1]      [,2]
#>   M 0.3269419 0.1635154
#>   R 0.2591308 0.1589568
#> 
#>    V15
#> y        [,1]      [,2]
#>   M 0.3409081 0.1928362
#>   R 0.2917323 0.2142258
#> 
#>    V16
#> y        [,1]      [,2]
#>   M 0.3869473 0.2129134
#>   R 0.3409400 0.2507794
#> 
#>    V17
#> y        [,1]      [,2]
#>   M 0.4242297 0.2423359
#>   R 0.3648415 0.2742038
#> 
#>    V18
#> y        [,1]      [,2]
#>   M 0.4750378 0.2564281
#>   R 0.4124569 0.2501830
#> 
#>    V19
#> y        [,1]      [,2]
#>   M 0.5523541 0.2525782
#>   R 0.4537400 0.2421448
#> 
#>    V2
#> y         [,1]       [,2]
#>   M 0.04152703 0.03123801
#>   R 0.03036923 0.02615068
#> 
#>    V20
#> y        [,1]      [,2]
#>   M 0.6250081 0.2534773
#>   R 0.5012169 0.2529609
#> 
#>    V21
#> y        [,1]      [,2]
#>   M 0.6804054 0.2494506
#>   R 0.5404938 0.2466701
#> 
#>    V22
#> y        [,1]      [,2]
#>   M 0.6749446 0.2498371
#>   R 0.5738277 0.2612811
#> 
#>    V23
#> y        [,1]      [,2]
#>   M 0.6732865 0.2654633
#>   R 0.6374185 0.2474285
#> 
#>    V24
#> y        [,1]      [,2]
#>   M 0.6997703 0.2454679
#>   R 0.6890262 0.2250570
#> 
#>    V25
#> y        [,1]      [,2]
#>   M 0.6962014 0.2405351
#>   R 0.6970031 0.2439968
#> 
#>    V26
#> y        [,1]      [,2]
#>   M 0.7233932 0.2302072
#>   R 0.7125200 0.2265733
#> 
#>    V27
#> y        [,1]      [,2]
#>   M 0.7301851 0.2668863
#>   R 0.6946338 0.2187736
#> 
#>    V28
#> y        [,1]      [,2]
#>   M 0.7065811 0.2745744
#>   R 0.6694046 0.2135123
#> 
#>    V29
#> y        [,1]      [,2]
#>   M 0.6383743 0.2504304
#>   R 0.6293815 0.2470967
#> 
#>    V3
#> y         [,1]       [,2]
#>   M 0.04895000 0.03806270
#>   R 0.03674462 0.03063795
#> 
#>    V30
#> y        [,1]      [,2]
#>   M 0.5568257 0.2095431
#>   R 0.5710338 0.2309240
#> 
#>    V31
#> y        [,1]      [,2]
#>   M 0.4462324 0.2168891
#>   R 0.5233938 0.1973785
#> 
#>    V32
#> y        [,1]      [,2]
#>   M 0.4025568 0.1957548
#>   R 0.4567492 0.2153948
#> 
#>    V33
#> y        [,1]      [,2]
#>   M 0.3777811 0.1877697
#>   R 0.4595031 0.2112573
#> 
#>    V34
#> y        [,1]      [,2]
#>   M 0.3429932 0.2000099
#>   R 0.4525985 0.2504447
#> 
#>    V35
#> y        [,1]      [,2]
#>   M 0.3120676 0.2380588
#>   R 0.4402185 0.2602866
#> 
#>    V36
#> y        [,1]      [,2]
#>   M 0.3095216 0.2509448
#>   R 0.4467985 0.2481707
#> 
#>    V37
#> y        [,1]      [,2]
#>   M 0.3201176 0.2252891
#>   R 0.4026477 0.2298454
#> 
#>    V38
#> y        [,1]      [,2]
#>   M 0.3200554 0.2027788
#>   R 0.3379523 0.2079753
#> 
#>    V39
#> y        [,1]      [,2]
#>   M 0.3241797 0.1684751
#>   R 0.3140738 0.2081152
#> 
#>    V4
#> y         [,1]       [,2]
#>   M 0.05958378 0.04170366
#>   R 0.04327538 0.03422337
#> 
#>    V40
#> y        [,1]      [,2]
#>   M 0.3044757 0.1430099
#>   R 0.3037200 0.1880619
#> 
#>    V41
#> y        [,1]      [,2]
#>   M 0.2841986 0.1437287
#>   R 0.2642600 0.1575307
#> 
#>    V42
#> y        [,1]      [,2]
#>   M 0.2809122 0.1434746
#>   R 0.2481831 0.1507060
#> 
#>    V43
#> y        [,1]      [,2]
#>   M 0.2584311 0.1308090
#>   R 0.2084985 0.1273211
#> 
#>    V44
#> y        [,1]      [,2]
#>   M 0.2267176 0.1385398
#>   R 0.1754200 0.1154875
#> 
#>    V45
#> y        [,1]       [,2]
#>   M 0.2118365 0.15590646
#>   R 0.1341785 0.09602863
#> 
#>    V46
#> y        [,1]       [,2]
#>   M 0.1697459 0.11493311
#>   R 0.1094538 0.07956321
#> 
#>    V47
#> y         [,1]       [,2]
#>   M 0.13426622 0.07935679
#>   R 0.08918615 0.05830296
#> 
#>    V48
#> y         [,1]       [,2]
#>   M 0.10365270 0.06472341
#>   R 0.06699385 0.04215228
#> 
#>    V49
#> y         [,1]       [,2]
#>   M 0.06056757 0.03560070
#>   R 0.03744923 0.02449127
#> 
#>    V5
#> y         [,1]       [,2]
#>   M 0.08217703 0.04765900
#>   R 0.06406769 0.05078625
#> 
#>    V50
#> y         [,1]       [,2]
#>   M 0.02116081 0.01237314
#>   R 0.01643077 0.01032582
#> 
#>    V51
#> y         [,1]       [,2]
#>   M 0.01860541 0.01142124
#>   R 0.01225538 0.00801368
#> 
#>    V52
#> y         [,1]        [,2]
#>   M 0.01423378 0.008024129
#>   R 0.01056462 0.006876192
#> 
#>    V53
#> y          [,1]        [,2]
#>   M 0.010720270 0.006362841
#>   R 0.009776923 0.006641060
#> 
#>    V54
#> y          [,1]        [,2]
#>   M 0.012233784 0.008654720
#>   R 0.009438462 0.005777183
#> 
#>    V55
#> y          [,1]        [,2]
#>   M 0.009060811 0.007334079
#>   R 0.008446154 0.005241245
#> 
#>    V56
#> y          [,1]        [,2]
#>   M 0.008756757 0.005591850
#>   R 0.007286154 0.004621041
#> 
#>    V57
#> y          [,1]        [,2]
#>   M 0.007448649 0.005228013
#>   R 0.007606154 0.005316510
#> 
#>    V58
#> y          [,1]        [,2]
#>   M 0.008698649 0.006810296
#>   R 0.006706154 0.004779444
#> 
#>    V59
#> y          [,1]        [,2]
#>   M 0.008081081 0.005919521
#>   R 0.006724615 0.004637383
#> 
#>    V6
#> y         [,1]       [,2]
#>   M 0.11097568 0.04806959
#>   R 0.09418462 0.06018832
#> 
#>    V60
#> y          [,1]        [,2]
#>   M 0.006412162 0.004137134
#>   R 0.005881538 0.003461525
#> 
#>    V7
#> y        [,1]       [,2]
#>   M 0.1322027 0.05982921
#>   R 0.1053446 0.05768441
#> 
#>    V8
#> y        [,1]       [,2]
#>   M 0.1554014 0.09015903
#>   R 0.1086815 0.06699557
#> 
#>    V9
#> y        [,1]       [,2]
#>   M 0.2112473 0.12204143
#>   R 0.1294400 0.08836828
#> 
#> classif.ce 
#>  0.3188406