Naive Bayes Classification Learner

Naive Bayes classification. Calls e1071::naiveBayes() from package e1071.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.naive_bayes")
lrn("classif.naive_bayes")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3learners, e1071

Parameters

Id	Type	Default	Range

eps	numeric	0	\((-\infty, \infty)\)

laplace	numeric	0	\([0, \infty)\)

threshold	numeric	0.001	\((-\infty, \infty)\)

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayes

Methods

Public methods

LearnerClassifNaiveBayes$new()

LearnerClassifNaiveBayes$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("e1071", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("classif.naive_bayes")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerClassifNaiveBayes:classif.naive_bayes>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, e1071
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor
#> * Properties: multiclass, twoclass
#> 
#> Naive Bayes Classifier for Discrete Predictors
#> 
#> Call:
#> naiveBayes.default(x = x, y = y)
#> 
#> A-priori probabilities:
#> y
#>         M         R 
#> 0.5323741 0.4676259 
#> 
#> Conditional probabilities:
#>    V1
#> y         [,1]       [,2]
#>   M 0.03159324 0.02264436
#>   R 0.02384769 0.01599523
#> 
#>    V10
#> y        [,1]      [,2]
#>   M 0.2420568 0.1441377
#>   R 0.1628246 0.1116953
#> 
#>    V11
#> y        [,1]      [,2]
#>   M 0.2732338 0.1234800
#>   R 0.1709677 0.1131478
#> 
#>    V12
#> y        [,1]      [,2]
#>   M 0.2814081 0.1110046
#>   R 0.1968123 0.1320291
#> 
#>    V13
#> y        [,1]      [,2]
#>   M 0.2920284 0.1074943
#>   R 0.2284323 0.1437249
#> 
#>    V14
#> y        [,1]      [,2]
#>   M 0.3102041 0.1555887
#>   R 0.2744800 0.1711339
#> 
#>    V15
#> y        [,1]      [,2]
#>   M 0.3289176 0.1919621
#>   R 0.3133385 0.2220090
#> 
#>    V16
#> y        [,1]      [,2]
#>   M 0.3888649 0.2129345
#>   R 0.3812815 0.2578295
#> 
#>    V17
#> y        [,1]      [,2]
#>   M 0.4298284 0.2339191
#>   R 0.4262892 0.2962364
#> 
#>    V18
#> y        [,1]      [,2]
#>   M 0.4792230 0.2521479
#>   R 0.4589354 0.2640556
#> 
#>    V19
#> y        [,1]      [,2]
#>   M 0.5677514 0.2512542
#>   R 0.4619415 0.2577712
#> 
#>    V2
#> y         [,1]       [,2]
#>   M 0.04416486 0.03388253
#>   R 0.03222154 0.02663366
#> 
#>    V20
#> y        [,1]      [,2]
#>   M 0.6469635 0.2502290
#>   R 0.5032554 0.2668585
#> 
#>    V21
#> y        [,1]      [,2]
#>   M 0.6887797 0.2413748
#>   R 0.5486846 0.2640956
#> 
#>    V22
#> y        [,1]      [,2]
#>   M 0.6852405 0.2360522
#>   R 0.5740123 0.2766851
#> 
#>    V23
#> y        [,1]      [,2]
#>   M 0.6899730 0.2530759
#>   R 0.5924985 0.2542719
#> 
#>    V24
#> y        [,1]      [,2]
#>   M 0.7043243 0.2375220
#>   R 0.6239554 0.2456568
#> 
#>    V25
#> y        [,1]      [,2]
#>   M 0.6862014 0.2312559
#>   R 0.6476108 0.2780198
#> 
#>    V26
#> y        [,1]      [,2]
#>   M 0.7048905 0.2328247
#>   R 0.6768708 0.2487159
#> 
#>    V27
#> y        [,1]      [,2]
#>   M 0.7119500 0.2672050
#>   R 0.6612769 0.2242181
#> 
#>    V28
#> y        [,1]      [,2]
#>   M 0.7063946 0.2638540
#>   R 0.6478662 0.1994863
#> 
#>    V29
#> y        [,1]      [,2]
#>   M 0.6473203 0.2458040
#>   R 0.6066615 0.2286503
#> 
#>    V3
#> y         [,1]       [,2]
#>   M 0.05023514 0.03908873
#>   R 0.03617385 0.03016417
#> 
#>    V30
#> y        [,1]      [,2]
#>   M 0.5839878 0.1997198
#>   R 0.5473831 0.2281882
#> 
#>    V31
#> y        [,1]      [,2]
#>   M 0.4669514 0.2130995
#>   R 0.5139200 0.1945606
#> 
#>    V32
#> y        [,1]      [,2]
#>   M 0.4101419 0.2045328
#>   R 0.4483538 0.2134295
#> 
#>    V33
#> y        [,1]      [,2]
#>   M 0.3802743 0.1909435
#>   R 0.4504600 0.2148139
#> 
#>    V34
#> y        [,1]      [,2]
#>   M 0.3579500 0.2057945
#>   R 0.4604631 0.2580194
#> 
#>    V35
#> y        [,1]      [,2]
#>   M 0.3426122 0.2440268
#>   R 0.4571323 0.2696266
#> 
#>    V36
#> y        [,1]      [,2]
#>   M 0.3185405 0.2612676
#>   R 0.4668769 0.2732720
#> 
#>    V37
#> y        [,1]      [,2]
#>   M 0.3232757 0.2375983
#>   R 0.4329615 0.2558516
#> 
#>    V38
#> y        [,1]      [,2]
#>   M 0.3444622 0.2134840
#>   R 0.3757615 0.2307133
#> 
#>    V39
#> y        [,1]      [,2]
#>   M 0.3366189 0.1904792
#>   R 0.3327769 0.2220780
#> 
#>    V4
#> y         [,1]       [,2]
#>   M 0.06202432 0.04432214
#>   R 0.04300308 0.03442376
#> 
#>    V40
#> y        [,1]      [,2]
#>   M 0.3082905 0.1625879
#>   R 0.3404169 0.1963534
#> 
#>    V41
#> y        [,1]      [,2]
#>   M 0.3072122 0.1690866
#>   R 0.3070585 0.1760232
#> 
#>    V42
#> y        [,1]      [,2]
#>   M 0.3069392 0.1641837
#>   R 0.2659323 0.1679813
#> 
#>    V43
#> y        [,1]      [,2]
#>   M 0.2713757 0.1321253
#>   R 0.2148600 0.1178370
#> 
#>    V44
#> y        [,1]       [,2]
#>   M 0.2439189 0.13159793
#>   R 0.1693800 0.08704486
#> 
#>    V45
#> y        [,1]      [,2]
#>   M 0.2429797 0.1702849
#>   R 0.1412969 0.0813411
#> 
#>    V46
#> y        [,1]       [,2]
#>   M 0.2049838 0.14944563
#>   R 0.1205262 0.09410163
#> 
#>    V47
#> y         [,1]       [,2]
#>   M 0.14953243 0.09392773
#>   R 0.09782154 0.07059624
#> 
#>    V48
#> y         [,1]       [,2]
#>   M 0.11124054 0.06868925
#>   R 0.07153385 0.05017188
#> 
#>    V49
#> y         [,1]       [,2]
#>   M 0.06363243 0.03872639
#>   R 0.04080154 0.03121400
#> 
#>    V5
#> y         [,1]       [,2]
#>   M 0.08801351 0.05432798
#>   R 0.06626308 0.05107537
#> 
#>    V50
#> y         [,1]       [,2]
#>   M 0.02270541 0.01511521
#>   R 0.01897692 0.01422650
#> 
#>    V51
#> y         [,1]        [,2]
#>   M 0.01852568 0.012149334
#>   R 0.01282462 0.008948831
#> 
#>    V52
#> y         [,1]        [,2]
#>   M 0.01573243 0.009604709
#>   R 0.01081692 0.006776189
#> 
#>    V53
#> y         [,1]        [,2]
#>   M 0.01184189 0.007774793
#>   R 0.01049231 0.006412934
#> 
#>    V54
#> y         [,1]        [,2]
#>   M 0.01195946 0.008718928
#>   R 0.00996000 0.005464653
#> 
#>    V55
#> y          [,1]        [,2]
#>   M 0.010014865 0.009238984
#>   R 0.008726154 0.005141871
#> 
#>    V56
#> y          [,1]        [,2]
#>   M 0.009408108 0.007054266
#>   R 0.007467692 0.004574662
#> 
#>    V57
#> y          [,1]        [,2]
#>   M 0.008200000 0.006339634
#>   R 0.007910769 0.005338250
#> 
#>    V58
#> y          [,1]        [,2]
#>   M 0.008902703 0.007371901
#>   R 0.006233846 0.004111047
#> 
#>    V59
#> y          [,1]        [,2]
#>   M 0.008322973 0.007222711
#>   R 0.006506154 0.004471048
#> 
#>    V6
#> y         [,1]       [,2]
#>   M 0.11019324 0.05268253
#>   R 0.09867077 0.07033397
#> 
#>    V60
#> y          [,1]        [,2]
#>   M 0.006818919 0.006335217
#>   R 0.005975385 0.003255478
#> 
#>    V7
#> y        [,1]       [,2]
#>   M 0.1250797 0.05968057
#>   R 0.1152138 0.06968515
#> 
#>    V8
#> y        [,1]       [,2]
#>   M 0.1471459 0.08804977
#>   R 0.1215554 0.08381520
#> 
#>    V9
#> y        [,1]      [,2]
#>   M 0.2117919 0.1277786
#>   R 0.1397246 0.1023128
#> 
#> classif.ce 
#>  0.2173913