Naive Bayes Classification Learner

Naive Bayes classification. Calls e1071::naiveBayes() from package e1071.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.naive_bayes")
lrn("classif.naive_bayes")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3learners, e1071

Parameters

Id	Type	Default	Range

eps	numeric	0	\((-\infty, \infty)\)

laplace	numeric	0	\([0, \infty)\)

threshold	numeric	0.001	\((-\infty, \infty)\)

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayes

Methods

Public methods

LearnerClassifNaiveBayes$new()

LearnerClassifNaiveBayes$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("e1071", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("classif.naive_bayes")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerClassifNaiveBayes:classif.naive_bayes>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, e1071
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor
#> * Properties: multiclass, twoclass
#> 
#> Naive Bayes Classifier for Discrete Predictors
#> 
#> Call:
#> naiveBayes.default(x = x, y = y)
#> 
#> A-priori probabilities:
#> y
#>         M         R 
#> 0.5323741 0.4676259 
#> 
#> Conditional probabilities:
#>    V1
#> y         [,1]       [,2]
#>   M 0.03600000 0.02830856
#>   R 0.02272154 0.01597110
#> 
#>    V10
#> y        [,1]      [,2]
#>   M 0.2487338 0.1339627
#>   R 0.1585138 0.1053588
#> 
#>    V11
#> y        [,1]      [,2]
#>   M 0.2908959 0.1297812
#>   R 0.1766077 0.1112859
#> 
#>    V12
#> y        [,1]      [,2]
#>   M 0.3086486 0.1281930
#>   R 0.1874462 0.1353777
#> 
#>    V13
#> y        [,1]      [,2]
#>   M 0.3254297 0.1411223
#>   R 0.2295231 0.1377307
#> 
#>    V14
#> y        [,1]      [,2]
#>   M 0.3397000 0.1719226
#>   R 0.2737631 0.1655535
#> 
#>    V15
#> y        [,1]      [,2]
#>   M 0.3634459 0.2022442
#>   R 0.3065754 0.2159388
#> 
#>    V16
#> y        [,1]      [,2]
#>   M 0.4208527 0.2154224
#>   R 0.3731938 0.2535482
#> 
#>    V17
#> y        [,1]      [,2]
#>   M 0.4543041 0.2400326
#>   R 0.4334123 0.2883129
#> 
#>    V18
#> y        [,1]      [,2]
#>   M 0.4878365 0.2573474
#>   R 0.4610262 0.2655775
#> 
#>    V19
#> y        [,1]      [,2]
#>   M 0.5712243 0.2450751
#>   R 0.4549969 0.2597537
#> 
#>    V2
#> y         [,1]       [,2]
#>   M 0.04730270 0.04034668
#>   R 0.03076615 0.02672007
#> 
#>    V20
#> y        [,1]      [,2]
#>   M 0.6321027 0.2542967
#>   R 0.4881538 0.2654051
#> 
#>    V21
#> y        [,1]      [,2]
#>   M 0.6656203 0.2539833
#>   R 0.5452000 0.2548954
#> 
#>    V22
#> y        [,1]      [,2]
#>   M 0.6722743 0.2559487
#>   R 0.5846262 0.2583370
#> 
#>    V23
#> y        [,1]      [,2]
#>   M 0.6756257 0.2599176
#>   R 0.6139723 0.2411046
#> 
#>    V24
#> y        [,1]      [,2]
#>   M 0.6816216 0.2545460
#>   R 0.6565769 0.2259674
#> 
#>    V25
#> y        [,1]      [,2]
#>   M 0.6771946 0.2381500
#>   R 0.6925246 0.2327587
#> 
#>    V26
#> y        [,1]      [,2]
#>   M 0.6906243 0.2495596
#>   R 0.7161031 0.2307044
#> 
#>    V27
#> y        [,1]      [,2]
#>   M 0.6836905 0.2773718
#>   R 0.6925692 0.2286364
#> 
#>    V28
#> y        [,1]      [,2]
#>   M 0.6718108 0.2690551
#>   R 0.6849554 0.2080270
#> 
#>    V29
#> y        [,1]      [,2]
#>   M 0.6163365 0.2613672
#>   R 0.6524400 0.2261379
#> 
#>    V3
#> y         [,1]       [,2]
#>   M 0.05258514 0.04882753
#>   R 0.03621385 0.03008432
#> 
#>    V30
#> y        [,1]      [,2]
#>   M 0.5611554 0.2286262
#>   R 0.5951462 0.2301385
#> 
#>    V31
#> y        [,1]      [,2]
#>   M 0.4733500 0.2224556
#>   R 0.5437138 0.2046520
#> 
#>    V32
#> y        [,1]      [,2]
#>   M 0.4188932 0.2070164
#>   R 0.4671338 0.2221197
#> 
#>    V33
#> y        [,1]      [,2]
#>   M 0.3870973 0.1951462
#>   R 0.4483415 0.2272859
#> 
#>    V34
#> y        [,1]      [,2]
#>   M 0.3559041 0.2017620
#>   R 0.4515908 0.2597925
#> 
#>    V35
#> y        [,1]      [,2]
#>   M 0.3263743 0.2322241
#>   R 0.4813477 0.2581959
#> 
#>    V36
#> y        [,1]      [,2]
#>   M 0.3090486 0.2272635
#>   R 0.4947108 0.2546924
#> 
#>    V37
#> y        [,1]      [,2]
#>   M 0.3131311 0.2127895
#>   R 0.4253554 0.2382752
#> 
#>    V38
#> y        [,1]      [,2]
#>   M 0.3277892 0.1808095
#>   R 0.3303769 0.2130491
#> 
#>    V39
#> y        [,1]      [,2]
#>   M 0.3288757 0.1683930
#>   R 0.2906169 0.2096892
#> 
#>    V4
#> y         [,1]       [,2]
#>   M 0.06971351 0.06161310
#>   R 0.04253385 0.03198846
#> 
#>    V40
#> y        [,1]      [,2]
#>   M 0.3044730 0.1617958
#>   R 0.3081969 0.1822925
#> 
#>    V41
#> y        [,1]      [,2]
#>   M 0.2990770 0.1629181
#>   R 0.2745662 0.1567462
#> 
#>    V42
#> y        [,1]      [,2]
#>   M 0.3056108 0.1621203
#>   R 0.2426046 0.1491485
#> 
#>    V43
#> y        [,1]      [,2]
#>   M 0.2823284 0.1387428
#>   R 0.2126169 0.1118267
#> 
#>    V44
#> y        [,1]       [,2]
#>   M 0.2487324 0.13959671
#>   R 0.1733354 0.08872881
#> 
#>    V45
#> y        [,1]       [,2]
#>   M 0.2420527 0.17002160
#>   R 0.1365631 0.08107176
#> 
#>    V46
#> y        [,1]       [,2]
#>   M 0.1991851 0.15472683
#>   R 0.1106200 0.08740801
#> 
#>    V47
#> y         [,1]       [,2]
#>   M 0.14480541 0.10120819
#>   R 0.09300462 0.06741345
#> 
#>    V48
#> y         [,1]       [,2]
#>   M 0.10742568 0.07144593
#>   R 0.07359538 0.05150681
#> 
#>    V49
#> y         [,1]       [,2]
#>   M 0.06332973 0.03821573
#>   R 0.04069538 0.03436357
#> 
#>    V5
#> y         [,1]       [,2]
#>   M 0.09299189 0.06614245
#>   R 0.05917538 0.04710075
#> 
#>    V50
#> y         [,1]       [,2]
#>   M 0.02267162 0.01481348
#>   R 0.01797077 0.01364552
#> 
#>    V51
#> y         [,1]        [,2]
#>   M 0.01952432 0.014720201
#>   R 0.01210308 0.007851611
#> 
#>    V52
#> y          [,1]        [,2]
#>   M 0.016817568 0.011762366
#>   R 0.009967692 0.005599166
#> 
#>    V53
#> y          [,1]        [,2]
#>   M 0.012990541 0.008176716
#>   R 0.009707692 0.006325526
#> 
#>    V54
#> y          [,1]        [,2]
#>   M 0.012843243 0.008221723
#>   R 0.009729231 0.005369670
#> 
#>    V55
#> y          [,1]        [,2]
#>   M 0.010345946 0.008458896
#>   R 0.008369231 0.005207954
#> 
#>    V56
#> y          [,1]        [,2]
#>   M 0.008824324 0.006151151
#>   R 0.007570769 0.004766115
#> 
#>    V57
#> y          [,1]        [,2]
#>   M 0.008164865 0.006378913
#>   R 0.008172308 0.005502911
#> 
#>    V58
#> y          [,1]        [,2]
#>   M 0.009479730 0.008283702
#>   R 0.006961538 0.004564732
#> 
#>    V59
#> y          [,1]        [,2]
#>   M 0.008589189 0.006382072
#>   R 0.006812308 0.005048314
#> 
#>    V6
#> y         [,1]       [,2]
#>   M 0.11244054 0.05487447
#>   R 0.08912923 0.06300852
#> 
#>    V60
#> y          [,1]        [,2]
#>   M 0.006979730 0.006383647
#>   R 0.005876923 0.003670779
#> 
#>    V7
#> y        [,1]       [,2]
#>   M 0.1317041 0.05971911
#>   R 0.1098862 0.06487213
#> 
#>    V8
#> y        [,1]       [,2]
#>   M 0.1474730 0.08026386
#>   R 0.1200062 0.08095063
#> 
#>    V9
#> y        [,1]       [,2]
#>   M 0.2091757 0.11208321
#>   R 0.1368569 0.09636783
#> 
#> classif.ce 
#>  0.3623188