Naive Bayes Classification Learner

Naive Bayes classification. Calls e1071::naiveBayes() from package e1071.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.naive_bayes")
lrn("classif.naive_bayes")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3learners, e1071

Parameters

Id	Type	Default	Range

eps	numeric	0	\((-\infty, \infty)\)

laplace	numeric	0	\([0, \infty)\)

threshold	numeric	0.001	\((-\infty, \infty)\)

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayes

Methods

Public methods

LearnerClassifNaiveBayes$new()

LearnerClassifNaiveBayes$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("e1071", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("classif.naive_bayes")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerClassifNaiveBayes:classif.naive_bayes>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, e1071
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor
#> * Properties: multiclass, twoclass
#> 
#> Naive Bayes Classifier for Discrete Predictors
#> 
#> Call:
#> naiveBayes.default(x = x, y = y)
#> 
#> A-priori probabilities:
#> y
#>         M         R 
#> 0.5323741 0.4676259 
#> 
#> Conditional probabilities:
#>    V1
#> y         [,1]       [,2]
#>   M 0.03211216 0.02646118
#>   R 0.02258615 0.01273566
#> 
#>    V10
#> y        [,1]     [,2]
#>   M 0.2559905 0.136201
#>   R 0.1645846 0.119452
#> 
#>    V11
#> y        [,1]      [,2]
#>   M 0.2940176 0.1230034
#>   R 0.1777354 0.1216786
#> 
#>    V12
#> y        [,1]      [,2]
#>   M 0.2960189 0.1188335
#>   R 0.1999031 0.1475704
#> 
#>    V13
#> y        [,1]      [,2]
#>   M 0.3010311 0.1261855
#>   R 0.2385677 0.1462754
#> 
#>    V14
#> y        [,1]      [,2]
#>   M 0.3015716 0.1637763
#>   R 0.2744292 0.1807053
#> 
#>    V15
#> y        [,1]      [,2]
#>   M 0.3060041 0.1911346
#>   R 0.3077000 0.2260937
#> 
#>    V16
#> y        [,1]      [,2]
#>   M 0.3493216 0.2093512
#>   R 0.3645400 0.2610860
#> 
#>    V17
#> y        [,1]      [,2]
#>   M 0.3806122 0.2324043
#>   R 0.4008615 0.2915949
#> 
#>    V18
#> y        [,1]      [,2]
#>   M 0.4320446 0.2465643
#>   R 0.4378492 0.2683712
#> 
#>    V19
#> y        [,1]      [,2]
#>   M 0.5132270 0.2444131
#>   R 0.4618846 0.2579694
#> 
#>    V2
#> y         [,1]       [,2]
#>   M 0.04249324 0.03655773
#>   R 0.03337846 0.02703920
#> 
#>    V20
#> y        [,1]      [,2]
#>   M 0.5993338 0.2551154
#>   R 0.4972369 0.2616590
#> 
#>    V21
#> y        [,1]      [,2]
#>   M 0.6486270 0.2566659
#>   R 0.5322492 0.2524621
#> 
#>    V22
#> y        [,1]      [,2]
#>   M 0.6545473 0.2513833
#>   R 0.5548692 0.2608728
#> 
#>    V23
#> y        [,1]      [,2]
#>   M 0.6679514 0.2577246
#>   R 0.6050262 0.2463647
#> 
#>    V24
#> y        [,1]      [,2]
#>   M 0.6895649 0.2518771
#>   R 0.6505292 0.2403723
#> 
#>    V25
#> y        [,1]      [,2]
#>   M 0.6963568 0.2428372
#>   R 0.6566585 0.2552699
#> 
#>    V26
#> y        [,1]      [,2]
#>   M 0.7328527 0.2325638
#>   R 0.6840492 0.2389274
#> 
#>    V27
#> y        [,1]      [,2]
#>   M 0.7383514 0.2683069
#>   R 0.6725877 0.2370179
#> 
#>    V28
#> y        [,1]      [,2]
#>   M 0.7290189 0.2663453
#>   R 0.6730062 0.2090686
#> 
#>    V29
#> y        [,1]      [,2]
#>   M 0.6636486 0.2510773
#>   R 0.6488262 0.2428132
#> 
#>    V3
#> y         [,1]       [,2]
#>   M 0.04681351 0.04421369
#>   R 0.03878615 0.03057077
#> 
#>    V30
#> y        [,1]      [,2]
#>   M 0.5927500 0.2055920
#>   R 0.5894108 0.2460983
#> 
#>    V31
#> y        [,1]      [,2]
#>   M 0.4902405 0.2074868
#>   R 0.5484369 0.2107259
#> 
#>    V32
#> y        [,1]      [,2]
#>   M 0.4272446 0.1965520
#>   R 0.4708862 0.2199134
#> 
#>    V33
#> y        [,1]      [,2]
#>   M 0.3997419 0.1886861
#>   R 0.4564354 0.2226831
#> 
#>    V34
#> y        [,1]      [,2]
#>   M 0.3677338 0.2044648
#>   R 0.4712462 0.2477581
#> 
#>    V35
#> y        [,1]      [,2]
#>   M 0.3294581 0.2478514
#>   R 0.4772815 0.2612378
#> 
#>    V36
#> y        [,1]      [,2]
#>   M 0.3278095 0.2592077
#>   R 0.4743892 0.2654924
#> 
#>    V37
#> y        [,1]      [,2]
#>   M 0.3367297 0.2415693
#>   R 0.4250477 0.2463619
#> 
#>    V38
#> y        [,1]      [,2]
#>   M 0.3364635 0.2125834
#>   R 0.3593400 0.2224475
#> 
#>    V39
#> y        [,1]      [,2]
#>   M 0.3336324 0.1805753
#>   R 0.3120246 0.2140145
#> 
#>    V4
#> y         [,1]       [,2]
#>   M 0.05931216 0.05404252
#>   R 0.04340462 0.03342860
#> 
#>    V40
#> y        [,1]      [,2]
#>   M 0.3028041 0.1589026
#>   R 0.3093123 0.1936789
#> 
#>    V41
#> y        [,1]      [,2]
#>   M 0.2680703 0.1501275
#>   R 0.2851185 0.1847801
#> 
#>    V42
#> y        [,1]      [,2]
#>   M 0.2787824 0.1413191
#>   R 0.2609969 0.1756992
#> 
#>    V43
#> y        [,1]      [,2]
#>   M 0.2689595 0.1261663
#>   R 0.2215246 0.1365506
#> 
#>    V44
#> y        [,1]      [,2]
#>   M 0.2482203 0.1390744
#>   R 0.1836862 0.1128775
#> 
#>    V45
#> y        [,1]       [,2]
#>   M 0.2307892 0.15658099
#>   R 0.1451738 0.09796111
#> 
#>    V46
#> y        [,1]       [,2]
#>   M 0.1807473 0.12739323
#>   R 0.1163446 0.09823192
#> 
#>    V47
#> y         [,1]       [,2]
#>   M 0.14091081 0.08759355
#>   R 0.09506615 0.06896941
#> 
#>    V48
#> y         [,1]       [,2]
#>   M 0.11035811 0.06813223
#>   R 0.06934154 0.04736968
#> 
#>    V49
#> y         [,1]       [,2]
#>   M 0.06098514 0.03557902
#>   R 0.03690154 0.03009637
#> 
#>    V5
#> y         [,1]       [,2]
#>   M 0.07674324 0.05784158
#>   R 0.06412923 0.04809227
#> 
#>    V50
#> y         [,1]       [,2]
#>   M 0.02015000 0.01269180
#>   R 0.01823538 0.01372907
#> 
#>    V51
#> y         [,1]        [,2]
#>   M 0.01837568 0.013376414
#>   R 0.01256000 0.009255367
#> 
#>    V52
#> y         [,1]        [,2]
#>   M 0.01523514 0.010706578
#>   R 0.01103077 0.007893152
#> 
#>    V53
#> y         [,1]        [,2]
#>   M 0.01102162 0.007453900
#>   R 0.01030769 0.006323155
#> 
#>    V54
#> y          [,1]        [,2]
#>   M 0.011516216 0.007954460
#>   R 0.008956923 0.005102204
#> 
#>    V55
#> y          [,1]        [,2]
#>   M 0.009610811 0.008736721
#>   R 0.009026154 0.005424691
#> 
#>    V56
#> y          [,1]        [,2]
#>   M 0.008589189 0.007073732
#>   R 0.008127692 0.005223580
#> 
#>    V57
#> y          [,1]        [,2]
#>   M 0.007770270 0.005973933
#>   R 0.007938462 0.005724582
#> 
#>    V58
#> y          [,1]        [,2]
#>   M 0.009136486 0.008077070
#>   R 0.006798462 0.004918825
#> 
#>    V59
#> y          [,1]        [,2]
#>   M 0.007929730 0.006406818
#>   R 0.006992308 0.004433724
#> 
#>    V6
#> y        [,1]       [,2]
#>   M 0.1061973 0.04778518
#>   R 0.1000892 0.06871395
#> 
#>    V60
#> y          [,1]        [,2]
#>   M 0.006275676 0.004796854
#>   R 0.006189231 0.003300811
#> 
#>    V7
#> y        [,1]       [,2]
#>   M 0.1262676 0.05768971
#>   R 0.1144415 0.06873639
#> 
#>    V8
#> y        [,1]       [,2]
#>   M 0.1463703 0.09585805
#>   R 0.1224108 0.08498743
#> 
#>    V9
#> y        [,1]      [,2]
#>   M 0.2151270 0.1312919
#>   R 0.1421246 0.1063278
#> 
#> classif.ce 
#>   0.173913