Naive Bayes Classification Learner

Naive Bayes classification. Calls e1071::naiveBayes() from package e1071.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.naive_bayes")
lrn("classif.naive_bayes")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3learners, e1071

Parameters

Id	Type	Default	Range

eps	numeric	0	\((-\infty, \infty)\)

laplace	numeric	0	\([0, \infty)\)

threshold	numeric	0.001	\((-\infty, \infty)\)

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayes

Methods

Public methods

LearnerClassifNaiveBayes$new()

LearnerClassifNaiveBayes$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("e1071", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("classif.naive_bayes")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerClassifNaiveBayes:classif.naive_bayes>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, e1071
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor
#> * Properties: multiclass, twoclass
#> 
#> Naive Bayes Classifier for Discrete Predictors
#> 
#> Call:
#> naiveBayes.default(x = x, y = y)
#> 
#> A-priori probabilities:
#> y
#>         M         R 
#> 0.5323741 0.4676259 
#> 
#> Conditional probabilities:
#>    V1
#> y         [,1]       [,2]
#>   M 0.03480541 0.02682549
#>   R 0.02160154 0.01421152
#> 
#>    V10
#> y        [,1]      [,2]
#>   M 0.2483973 0.1412391
#>   R 0.1605938 0.1130638
#> 
#>    V11
#> y        [,1]      [,2]
#>   M 0.2934541 0.1253682
#>   R 0.1709508 0.1087527
#> 
#>    V12
#> y        [,1]      [,2]
#>   M 0.3074230 0.1208811
#>   R 0.1845385 0.1332908
#> 
#>    V13
#> y        [,1]      [,2]
#>   M 0.3262041 0.1302433
#>   R 0.2247462 0.1361198
#> 
#>    V14
#> y        [,1]      [,2]
#>   M 0.3343378 0.1666415
#>   R 0.2705262 0.1690585
#> 
#>    V15
#> y        [,1]      [,2]
#>   M 0.3403419 0.1941862
#>   R 0.3073446 0.2250075
#> 
#>    V16
#> y        [,1]      [,2]
#>   M 0.3831419 0.2139784
#>   R 0.3639031 0.2676098
#> 
#>    V17
#> y        [,1]      [,2]
#>   M 0.4198297 0.2451442
#>   R 0.4066108 0.2990857
#> 
#>    V18
#> y        [,1]      [,2]
#>   M 0.4633865 0.2566164
#>   R 0.4373031 0.2794899
#> 
#>    V19
#> y        [,1]      [,2]
#>   M 0.5383162 0.2602064
#>   R 0.4326754 0.2676141
#> 
#>    V2
#> y         [,1]       [,2]
#>   M 0.04795946 0.03971273
#>   R 0.03137692 0.02695484
#> 
#>    V20
#> y        [,1]      [,2]
#>   M 0.6148095 0.2518921
#>   R 0.4579615 0.2473310
#> 
#>    V21
#> y        [,1]      [,2]
#>   M 0.6794243 0.2450771
#>   R 0.5174662 0.2488511
#> 
#>    V22
#> y        [,1]      [,2]
#>   M 0.6786676 0.2371034
#>   R 0.5632492 0.2738436
#> 
#>    V23
#> y        [,1]      [,2]
#>   M 0.6812527 0.2600112
#>   R 0.6191369 0.2533875
#> 
#>    V24
#> y        [,1]      [,2]
#>   M 0.6945216 0.2484351
#>   R 0.6585108 0.2395248
#> 
#>    V25
#> y        [,1]      [,2]
#>   M 0.6766662 0.2421630
#>   R 0.6787154 0.2447629
#> 
#>    V26
#> y        [,1]      [,2]
#>   M 0.7045473 0.2281226
#>   R 0.6965692 0.2425409
#> 
#>    V27
#> y        [,1]      [,2]
#>   M 0.7154595 0.2568411
#>   R 0.6983708 0.2091980
#> 
#>    V28
#> y        [,1]      [,2]
#>   M 0.7177649 0.2558776
#>   R 0.6868508 0.2055336
#> 
#>    V29
#> y        [,1]      [,2]
#>   M 0.6576243 0.2359663
#>   R 0.6402677 0.2257929
#> 
#>    V3
#> y         [,1]       [,2]
#>   M 0.05641757 0.04931509
#>   R 0.03590308 0.03068199
#> 
#>    V30
#> y        [,1]      [,2]
#>   M 0.5703189 0.2009887
#>   R 0.5854969 0.2250826
#> 
#>    V31
#> y        [,1]      [,2]
#>   M 0.4663378 0.2137007
#>   R 0.5475923 0.1958865
#> 
#>    V32
#> y       [,1]      [,2]
#>   M 0.407173 0.2149589
#>   R 0.476620 0.2020529
#> 
#>    V33
#> y        [,1]      [,2]
#>   M 0.3883959 0.2011583
#>   R 0.4544662 0.2217938
#> 
#>    V34
#> y        [,1]      [,2]
#>   M 0.3590216 0.2087780
#>   R 0.4514677 0.2591795
#> 
#>    V35
#> y        [,1]      [,2]
#>   M 0.3271703 0.2393558
#>   R 0.4767138 0.2710045
#> 
#>    V36
#> y        [,1]      [,2]
#>   M 0.3054757 0.2392878
#>   R 0.4712600 0.2714161
#> 
#>    V37
#> y        [,1]     [,2]
#>   M 0.2912419 0.212572
#>   R 0.4071892 0.248540
#> 
#>    V38
#> y        [,1]      [,2]
#>   M 0.3086324 0.1958072
#>   R 0.3330354 0.2219089
#> 
#>    V39
#> y        [,1]      [,2]
#>   M 0.3238649 0.1757909
#>   R 0.2891354 0.1934212
#> 
#>    V4
#> y         [,1]       [,2]
#>   M 0.07110000 0.06154191
#>   R 0.03994462 0.02909868
#> 
#>    V40
#> y        [,1]      [,2]
#>   M 0.3010446 0.1581729
#>   R 0.3069831 0.1720886
#> 
#>    V41
#> y        [,1]      [,2]
#>   M 0.2851081 0.1693641
#>   R 0.2803308 0.1633555
#> 
#>    V42
#> y        [,1]      [,2]
#>   M 0.2948662 0.1662927
#>   R 0.2531692 0.1546926
#> 
#>    V43
#> y        [,1]      [,2]
#>   M 0.2688689 0.1304758
#>   R 0.2168708 0.1180807
#> 
#>    V44
#> y        [,1]       [,2]
#>   M 0.2326554 0.13405153
#>   R 0.1752785 0.09314763
#> 
#>    V45
#> y        [,1]       [,2]
#>   M 0.2280919 0.16500171
#>   R 0.1314923 0.07383891
#> 
#>    V46
#> y         [,1]       [,2]
#>   M 0.18158378 0.13082483
#>   R 0.09935846 0.06146171
#> 
#>    V47
#> y         [,1]      [,2]
#>   M 0.13819324 0.0754807
#>   R 0.08463846 0.0493830
#> 
#>    V48
#> y         [,1]       [,2]
#>   M 0.10818243 0.05834892
#>   R 0.06734308 0.04168526
#> 
#>    V49
#> y         [,1]       [,2]
#>   M 0.06289595 0.03284159
#>   R 0.03589077 0.02575186
#> 
#>    V5
#> y         [,1]       [,2]
#>   M 0.09085135 0.06593174
#>   R 0.06041538 0.04360524
#> 
#>    V50
#> y         [,1]       [,2]
#>   M 0.02186216 0.01257010
#>   R 0.01635077 0.01052053
#> 
#>    V51
#> y         [,1]        [,2]
#>   M 0.01896892 0.010394927
#>   R 0.01223692 0.007541853
#> 
#>    V52
#> y          [,1]        [,2]
#>   M 0.015524324 0.009554828
#>   R 0.009898462 0.006548413
#> 
#>    V53
#> y          [,1]        [,2]
#>   M 0.011905405 0.007221238
#>   R 0.009443077 0.005874653
#> 
#>    V54
#> y          [,1]        [,2]
#>   M 0.012305405 0.008825460
#>   R 0.009367692 0.004899366
#> 
#>    V55
#> y          [,1]        [,2]
#>   M 0.009597297 0.007926615
#>   R 0.007478462 0.004483563
#> 
#>    V56
#> y          [,1]        [,2]
#>   M 0.008916216 0.006295531
#>   R 0.007173846 0.004657530
#> 
#>    V57
#> y          [,1]        [,2]
#>   M 0.007755405 0.005832529
#>   R 0.007478462 0.005374724
#> 
#>    V58
#> y          [,1]        [,2]
#>   M 0.009197297 0.007156910
#>   R 0.006358462 0.004316389
#> 
#>    V59
#> y          [,1]        [,2]
#>   M 0.009270270 0.007371227
#>   R 0.007066154 0.005003101
#> 
#>    V6
#> y         [,1]       [,2]
#>   M 0.11876081 0.05588745
#>   R 0.09832154 0.06475974
#> 
#>    V60
#> y          [,1]        [,2]
#>   M 0.007256757 0.006477296
#>   R 0.006018462 0.003724660
#> 
#>    V7
#> y        [,1]       [,2]
#>   M 0.1295784 0.05672309
#>   R 0.1127154 0.06460947
#> 
#>    V8
#> y        [,1]       [,2]
#>   M 0.1503270 0.08273623
#>   R 0.1109431 0.07360503
#> 
#>    V9
#> y        [,1]      [,2]
#>   M 0.2082527 0.1175178
#>   R 0.1381046 0.1027497
#> 
#> classif.ce 
#>  0.2898551