Naive Bayes Classification Learner

Naive Bayes classification. Calls e1071::naiveBayes() from package e1071.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.naive_bayes")
lrn("classif.naive_bayes")

Meta Information

Task type: “classif”

Predict Types: “response”, “prob”

Feature Types: “logical”, “integer”, “numeric”, “factor”

Required Packages: mlr3, mlr3learners, e1071

Parameters

Id	Type	Default	Range

eps	numeric	0	\((-\infty, \infty)\)

laplace	numeric	0	\([0, \infty)\)

threshold	numeric	0.001	\((-\infty, \infty)\)

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifNaiveBayes

Methods

Public methods

LearnerClassifNaiveBayes$new()

LearnerClassifNaiveBayes$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifNaiveBayes$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifNaiveBayes$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("e1071", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("classif.naive_bayes")
print(learner)

# Define a Task
task = tsk("sonar")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerClassifNaiveBayes:classif.naive_bayes>: Naive Bayes
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, e1071
#> * Predict Types:  [response], prob
#> * Feature Types: logical, integer, numeric, factor
#> * Properties: multiclass, twoclass
#> 
#> Naive Bayes Classifier for Discrete Predictors
#> 
#> Call:
#> naiveBayes.default(x = x, y = y)
#> 
#> A-priori probabilities:
#> y
#>         M         R 
#> 0.5323741 0.4676259 
#> 
#> Conditional probabilities:
#>    V1
#> y         [,1]       [,2]
#>   M 0.03698108 0.02839106
#>   R 0.02341538 0.01493843
#> 
#>    V10
#> y        [,1]      [,2]
#>   M 0.2481500 0.1566155
#>   R 0.1751154 0.1247372
#> 
#>    V11
#> y        [,1]      [,2]
#>   M 0.2896757 0.1400860
#>   R 0.1825123 0.1245659
#> 
#>    V12
#> y        [,1]      [,2]
#>   M 0.2930581 0.1299190
#>   R 0.2041215 0.1468243
#> 
#>    V13
#> y        [,1]      [,2]
#>   M 0.3062622 0.1364508
#>   R 0.2364585 0.1403513
#> 
#>    V14
#> y        [,1]      [,2]
#>   M 0.3257622 0.1608516
#>   R 0.2712723 0.1704194
#> 
#>    V15
#> y        [,1]      [,2]
#>   M 0.3465419 0.1962179
#>   R 0.3053508 0.2189969
#> 
#>    V16
#> y        [,1]      [,2]
#>   M 0.3974784 0.2144708
#>   R 0.3798015 0.2567753
#> 
#>    V17
#> y        [,1]      [,2]
#>   M 0.4312284 0.2371748
#>   R 0.4268631 0.3053064
#> 
#>    V18
#> y        [,1]      [,2]
#>   M 0.4749068 0.2551768
#>   R 0.4630277 0.2664581
#> 
#>    V19
#> y        [,1]      [,2]
#>   M 0.5578068 0.2492121
#>   R 0.4682477 0.2566260
#> 
#>    V2
#> y         [,1]       [,2]
#>   M 0.04718514 0.04185452
#>   R 0.03458615 0.02604459
#> 
#>    V20
#> y        [,1]      [,2]
#>   M 0.6420541 0.2427204
#>   R 0.4976062 0.2655010
#> 
#>    V21
#> y        [,1]      [,2]
#>   M 0.6949216 0.2374734
#>   R 0.5407354 0.2443076
#> 
#>    V22
#> y        [,1]      [,2]
#>   M 0.6926243 0.2378974
#>   R 0.5369800 0.2629888
#> 
#>    V23
#> y        [,1]      [,2]
#>   M 0.6885662 0.2606089
#>   R 0.5747477 0.2355679
#> 
#>    V24
#> y        [,1]      [,2]
#>   M 0.6869838 0.2554237
#>   R 0.6374138 0.2293929
#> 
#>    V25
#> y        [,1]      [,2]
#>   M 0.6624905 0.2509054
#>   R 0.6517769 0.2456621
#> 
#>    V26
#> y        [,1]      [,2]
#>   M 0.6806243 0.2499543
#>   R 0.6639292 0.2439895
#> 
#>    V27
#> y        [,1]      [,2]
#>   M 0.6837230 0.2786877
#>   R 0.6601154 0.2199334
#> 
#>    V28
#> y        [,1]      [,2]
#>   M 0.6884784 0.2652515
#>   R 0.6587062 0.1947668
#> 
#>    V29
#> y        [,1]      [,2]
#>   M 0.6381230 0.2344888
#>   R 0.6175138 0.2552503
#> 
#>    V3
#> y         [,1]       [,2]
#>   M 0.05273514 0.04844836
#>   R 0.04220923 0.03210370
#> 
#>    V30
#> y        [,1]      [,2]
#>   M 0.5568041 0.1981203
#>   R 0.5645615 0.2540082
#> 
#>    V31
#> y        [,1]      [,2]
#>   M 0.4471514 0.2049502
#>   R 0.5380631 0.2183655
#> 
#>    V32
#> y        [,1]      [,2]
#>   M 0.3985176 0.2018347
#>   R 0.4751569 0.2230354
#> 
#>    V33
#> y        [,1]      [,2]
#>   M 0.3824473 0.2017697
#>   R 0.4662938 0.2247103
#> 
#>    V34
#> y        [,1]      [,2]
#>   M 0.3770365 0.2110556
#>   R 0.4654015 0.2502735
#> 
#>    V35
#> y        [,1]      [,2]
#>   M 0.3689932 0.2540851
#>   R 0.4716554 0.2507160
#> 
#>    V36
#> y        [,1]      [,2]
#>   M 0.3438811 0.2547847
#>   R 0.4726769 0.2612290
#> 
#>    V37
#> y        [,1]      [,2]
#>   M 0.3234797 0.2385631
#>   R 0.4448954 0.2466739
#> 
#>    V38
#> y        [,1]      [,2]
#>   M 0.3366189 0.2088987
#>   R 0.3782815 0.2310386
#> 
#>    V39
#> y        [,1]      [,2]
#>   M 0.3380730 0.1902514
#>   R 0.3357508 0.2252771
#> 
#>    V4
#> y         [,1]       [,2]
#>   M 0.07009865 0.06127744
#>   R 0.04475692 0.03462266
#> 
#>    V40
#> y        [,1]      [,2]
#>   M 0.3070284 0.1569417
#>   R 0.3395492 0.1881423
#> 
#>    V41
#> y        [,1]      [,2]
#>   M 0.2968622 0.1625307
#>   R 0.3041862 0.1749181
#> 
#>    V42
#> y        [,1]      [,2]
#>   M 0.3074500 0.1715455
#>   R 0.2698938 0.1632931
#> 
#>    V43
#> y        [,1]      [,2]
#>   M 0.2721676 0.1341128
#>   R 0.2196923 0.1209920
#> 
#>    V44
#> y        [,1]       [,2]
#>   M 0.2388230 0.13323238
#>   R 0.1776954 0.09300673
#> 
#>    V45
#> y        [,1]       [,2]
#>   M 0.2439851 0.16523296
#>   R 0.1500154 0.08758075
#> 
#>    V46
#> y        [,1]       [,2]
#>   M 0.2008216 0.14685052
#>   R 0.1243200 0.08947079
#> 
#>    V47
#> y         [,1]       [,2]
#>   M 0.14404730 0.08694714
#>   R 0.09551231 0.06797832
#> 
#>    V48
#> y         [,1]       [,2]
#>   M 0.10373378 0.06050533
#>   R 0.07040462 0.05073246
#> 
#>    V49
#> y         [,1]       [,2]
#>   M 0.06147432 0.03270600
#>   R 0.04072923 0.03243654
#> 
#>    V5
#> y         [,1]       [,2]
#>   M 0.08970000 0.06259742
#>   R 0.06382462 0.05227831
#> 
#>    V50
#> y         [,1]       [,2]
#>   M 0.02282838 0.01193710
#>   R 0.01854308 0.01371346
#> 
#>    V51
#> y         [,1]        [,2]
#>   M 0.01924730 0.013360935
#>   R 0.01222308 0.008616677
#> 
#>    V52
#> y         [,1]        [,2]
#>   M 0.01707973 0.012157663
#>   R 0.01009385 0.006430345
#> 
#>    V53
#> y          [,1]        [,2]
#>   M 0.012297297 0.007916308
#>   R 0.009086154 0.005932508
#> 
#>    V54
#> y          [,1]        [,2]
#>   M 0.011904054 0.008567179
#>   R 0.009806154 0.005574492
#> 
#>    V55
#> y          [,1]        [,2]
#>   M 0.010366216 0.008517866
#>   R 0.008770769 0.005182534
#> 
#>    V56
#> y          [,1]        [,2]
#>   M 0.008435135 0.005242514
#>   R 0.007618462 0.005087512
#> 
#>    V57
#> y          [,1]        [,2]
#>   M 0.007547297 0.005305444
#>   R 0.008423077 0.005661183
#> 
#>    V58
#> y          [,1]        [,2]
#>   M 0.009683784 0.006766862
#>   R 0.007055385 0.004799936
#> 
#>    V59
#> y          [,1]        [,2]
#>   M 0.009420270 0.007423564
#>   R 0.007216923 0.004305105
#> 
#>    V6
#> y        [,1]       [,2]
#>   M 0.1133122 0.05937453
#>   R 0.1009954 0.07409183
#> 
#>    V60
#> y          [,1]        [,2]
#>   M 0.007509459 0.006781970
#>   R 0.006170769 0.003434554
#> 
#>    V7
#> y        [,1]       [,2]
#>   M 0.1300865 0.06326215
#>   R 0.1177415 0.07197998
#> 
#>    V8
#> y        [,1]       [,2]
#>   M 0.1550311 0.09305271
#>   R 0.1234431 0.08897038
#> 
#>    V9
#> y        [,1]      [,2]
#>   M 0.2114392 0.1313881
#>   R 0.1514723 0.1135924
#> 
#> classif.ce 
#>  0.2898551