Extreme Gradient Boosting Classification Learner

eXtreme Gradient Boosting classification. Calls xgboost::xgb.train() from package xgboost.

Custom mlr3 defaults

nrounds:

Actual default: no default.

Adjusted default: 1.

Reason for change: Without a default construction of the learner would error. Just setting a nonsense default to workaround this. nrounds needs to be tuned by the user.

nthread:

Actual value: Undefined, triggering auto-detection of the number of CPUs.

Adjusted value: 1.

Reason for change: Conflicting with parallelization via future.

verbose:

Actual default: 1.

Adjusted default: 0.

Reason for change: Reduce verbosity.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("classif.xgboost")
lrn("classif.xgboost")

References

Chen, Tianqi, Guestrin, Carlos (2016). “Xgboost: A scalable tree boosting system.” In Proceedings of the 22nd ACM SIGKDD Conference on Knowledge Discovery and Data Mining, 785--794. ACM. doi: 10.1145/2939672.2939785 .

See also

Super classes

mlr3::Learner -> mlr3::LearnerClassif -> LearnerClassifXgboost

Methods

Public methods

LearnerClassifXgboost$new()

LearnerClassifXgboost$importance()

LearnerClassifXgboost$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerClassifXgboost$new()

Method importance()

The importance scores are calculated with xgboost::xgb.importance().

Usage

LearnerClassifXgboost$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerClassifXgboost$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("xgboost")) {
  learner = mlr3::lrn("classif.xgboost")
  print(learner)

  # available parameters:
learner$param_set$ids()
}

#> Loading required namespace: xgboost

#> <LearnerClassifXgboost:classif.xgboost>
#> * Model: -
#> * Parameters: nrounds=1, nthread=1, verbose=0
#> * Packages: xgboost
#> * Predict Type: response
#> * Feature types: logical, integer, numeric
#> * Properties: importance, missings, multiclass, twoclass, weights

#>  [1] "alpha"                   "approxcontrib"          
#>  [3] "base_score"              "booster"                
#>  [5] "callbacks"               "colsample_bylevel"      
#>  [7] "colsample_bynode"        "colsample_bytree"       
#>  [9] "early_stopping_rounds"   "eta"                    
#> [11] "eval_metric"             "feature_selector"       
#> [13] "feval"                   "gamma"                  
#> [15] "grow_policy"             "interaction_constraints"
#> [17] "lambda"                  "lambda_bias"            
#> [19] "max_bin"                 "max_delta_step"         
#> [21] "max_depth"               "max_leaves"             
#> [23] "maximize"                "min_child_weight"       
#> [25] "missing"                 "monotone_constraints"   
#> [27] "normalize_type"          "nrounds"                
#> [29] "nthread"                 "ntreelimit"             
#> [31] "num_parallel_tree"       "objective"              
#> [33] "one_drop"                "outputmargin"           
#> [35] "predcontrib"             "predictor"              
#> [37] "predinteraction"         "predleaf"               
#> [39] "print_every_n"           "rate_drop"              
#> [41] "refresh_leaf"            "reshape"                
#> [43] "sample_type"             "save_name"              
#> [45] "save_period"             "scale_pos_weight"       
#> [47] "sketch_eps"              "skip_drop"              
#> [49] "subsample"               "top_k"                  
#> [51] "training"                "tree_method"            
#> [53] "tweedie_variance_power"  "updater"                
#> [55] "verbose"                 "watchlist"              
#> [57] "xgb_model"