Kriging Regression Learner

Kriging regression. Calls DiceKriging::km() from package DiceKriging.

The predict type hyperparameter "type" defaults to "SK" (simple kriging).

The additional hyperparameter nugget.stability is used to overwrite the hyperparameter nugget with nugget.stability * var(y) before training to improve the numerical stability. We recommend a value of 1e-8.

The additional hyperparameter jitter can be set to add N(0, [jitter])-distributed noise to the data before prediction to avoid perfect interpolation. We recommend a value of 1e-12.

Dictionary

This mlr3::Learner can be instantiated via the dictionary mlr3::mlr_learners or with the associated sugar function mlr3::lrn():

mlr_learners$get("regr.km")
lrn("regr.km")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3learners, DiceKriging

Parameters

Id	Type	Default	Levels	Range

bias.correct	logical	FALSE	TRUE, FALSE	-

checkNames	logical	TRUE	TRUE, FALSE	-

coef.cov	untyped	NULL		-

coef.trend	untyped	NULL		-

coef.var	untyped	NULL		-

control	untyped	NULL		-

cov.compute	logical	TRUE	TRUE, FALSE	-

covtype	character	matern5_2	gauss, matern5_2, matern3_2, exp, powexp	-

estim.method	character	MLE	MLE, LOO	-

gr	logical	TRUE	TRUE, FALSE	-

iso	logical	FALSE	TRUE, FALSE	-

jitter	numeric	0		\([0, \infty)\)

kernel	untyped	NULL		-

knots	untyped	NULL		-

light.return	logical	FALSE	TRUE, FALSE	-

lower	untyped	NULL		-

multistart	integer	1		\((-\infty, \infty)\)

noise.var	untyped	NULL		-

nugget	numeric	-		\((-\infty, \infty)\)

nugget.estim	logical	FALSE	TRUE, FALSE	-

nugget.stability	numeric	0		\([0, \infty)\)

optim.method	character	BFGS	BFGS, gen	-

parinit	untyped	NULL		-

penalty	untyped	NULL		-

scaling	logical	FALSE	TRUE, FALSE	-

se.compute	logical	TRUE	TRUE, FALSE	-

type	character	SK	SK, UK	-

upper	untyped	NULL		-

References

Roustant O, Ginsbourger D, Deville Y (2012). “DiceKriging, DiceOptim: Two R Packages for the Analysis of Computer Experiments by Kriging-Based Metamodeling and Optimization.” Journal of Statistical Software, 51(1), 1–55. doi:10.18637/jss.v051.i01 .

See also

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrKM

Methods

Public methods

LearnerRegrKM$new()

LearnerRegrKM$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$configure()

mlr3::Learner$encapsulate()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$selected_features()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrKM$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrKM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("DiceKriging", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("regr.km")
print(learner)

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerRegrKM:regr.km>: Kriging
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, DiceKriging
#> * Predict Types:  [response], se
#> * Feature Types: logical, integer, numeric
#> * Properties: -
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : BFGS 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern5_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 
#>   - parameters upper bounds :  2 14 8 786.6 2.92 4 538 11.44 2 7.822 
#>   - best initial criterion value(s) :  -59.83873 
#> 
#> N = 10, M = 5 machine precision = 2.22045e-16
#> At X0, 0 variables are exactly at the bounds
#> At iterate     0  f=       59.839  |proj g|=       3.6917
#> At iterate     1  f =        59.08  |proj g|=        1.5977
#> At iterate     2  f =       58.557  |proj g|=        1.2987
#> At iterate     3  f =       58.172  |proj g|=       0.34076
#> At iterate     4  f =       58.065  |proj g|=       0.36721
#> At iterate     5  f =       57.961  |proj g|=       0.26228
#> At iterate     6  f =        57.94  |proj g|=       0.15296
#> At iterate     7  f =       57.923  |proj g|=       0.10987
#> At iterate     8  f =       57.909  |proj g|=        0.1208
#> At iterate     9  f =       57.894  |proj g|=      0.092995
#> At iterate    10  f =       57.889  |proj g|=      0.045872
#> At iterate    11  f =       57.886  |proj g|=      0.059014
#> At iterate    12  f =        57.88  |proj g|=      0.058636
#> At iterate    13  f =       57.842  |proj g|=       0.15853
#> At iterate    14  f =       57.822  |proj g|=       0.11076
#> At iterate    15  f =       57.778  |proj g|=      0.045557
#> At iterate    16  f =       57.767  |proj g|=      0.036796
#> At iterate    17  f =       57.763  |proj g|=      0.034535
#> At iterate    18  f =        57.76  |proj g|=      0.033928
#> At iterate    19  f =       57.751  |proj g|=       0.06557
#> At iterate    20  f =       57.722  |proj g|=       0.13279
#> At iterate    21  f =       57.662  |proj g|=       0.15772
#> At iterate    22  f =        57.63  |proj g|=       0.64202
#> At iterate    23  f =       57.582  |proj g|=       0.10488
#> At iterate    24  f =       57.577  |proj g|=      0.046736
#> At iterate    25  f =       57.576  |proj g|=      0.024196
#> At iterate    26  f =       57.575  |proj g|=      0.017771
#> At iterate    27  f =       57.574  |proj g|=      0.037951
#> At iterate    28  f =       57.573  |proj g|=      0.081022
#> At iterate    29  f =       57.568  |proj g|=       0.14521
#> At iterate    30  f =       57.556  |proj g|=       0.23985
#> At iterate    31  f =       57.523  |proj g|=       0.36407
#> At iterate    32  f =       56.338  |proj g|=       0.61166
#> At iterate    33  f =       56.305  |proj g|=       0.42734
#> At iterate    34  f =       56.296  |proj g|=       0.38756
#> At iterate    35  f =       56.277  |proj g|=       0.10113
#> At iterate    36  f =       56.275  |proj g|=      0.088396
#> At iterate    37  f =       56.256  |proj g|=      0.014256
#> At iterate    38  f =        56.25  |proj g|=      0.022188
#> At iterate    39  f =       56.246  |proj g|=     0.0024186
#> At iterate    40  f =       56.246  |proj g|=     0.0012488
#> At iterate    41  f =       56.246  |proj g|=     0.0064794
#> At iterate    42  f =       56.246  |proj g|=      0.003192
#> At iterate    43  f =       56.246  |proj g|=     0.0012549
#> At iterate    44  f =       56.246  |proj g|=     0.0012971
#> At iterate    45  f =       56.246  |proj g|=     0.0012774
#> At iterate    46  f =       56.246  |proj g|=     0.0017957
#> At iterate    47  f =       56.246  |proj g|=     0.0012777
#> At iterate    48  f =       56.246  |proj g|=     0.0018004
#> At iterate    49  f =       56.246  |proj g|=      0.001278
#> At iterate    50  f =       56.246  |proj g|=     0.0013864
#> At iterate    51  f =       56.245  |proj g|=      0.033919
#> At iterate    52  f =       56.244  |proj g|=      0.086489
#> At iterate    53  f =        56.24  |proj g|=       0.19431
#> At iterate    54  f =       56.239  |proj g|=       0.18843
#> At iterate    55  f =       56.236  |proj g|=      0.056116
#> At iterate    56  f =       56.236  |proj g|=      0.002153
#> At iterate    57  f =       56.236  |proj g|=    0.00052824
#> At iterate    58  f =       56.236  |proj g|=    0.00052724
#> 
#> iterations 58
#> function evaluations 76
#> segments explored during Cauchy searches 58
#> BFGS updates skipped 0
#> active bounds at final generalized Cauchy point 5
#> norm of the final projected gradient 0.000527239
#> final function value 56.2361
#> 
#> F = 56.2361
#> final  value 56.236150 
#> converged
#> 
#> Call:
#> DiceKriging::km(design = data, response = truth, control = pv$control)
#> 
#> Trend  coeff.:
#>                Estimate
#>  (Intercept)    18.7597
#> 
#> Covar. type  : matern5_2 
#> Covar. coeff.:
#>                Estimate
#>    theta(am)     2.0000
#>  theta(carb)     7.1634
#>   theta(cyl)     8.0000
#>  theta(disp)   747.3931
#>  theta(drat)     2.9200
#>  theta(gear)     4.0000
#>    theta(hp)    33.6988
#>  theta(qsec)     1.2322
#>    theta(vs)     0.0000
#>    theta(wt)     7.8076
#> 
#> Variance estimate: 29.57219
#> regr.mse 
#> 17.04289