Kriging Regression Learner

Kriging regression. Calls DiceKriging::km() from package DiceKriging.

The predict type hyperparameter "type" defaults to "sk" (simple kriging).

The additional hyperparameter nugget.stability is used to overwrite the hyperparameter nugget with nugget.stability * var(y) before training to improve the numerical stability. We recommend a value of 1e-8.

The additional hyperparameter jitter can be set to add N(0, [jitter])-distributed noise to the data before prediction to avoid perfect interpolation. We recommend a value of 1e-12.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.km")
lrn("regr.km")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3learners, DiceKriging

Parameters

Id	Type	Default	Levels	Range

bias.correct	logical	FALSE	TRUE, FALSE	-

checkNames	logical	TRUE	TRUE, FALSE	-

coef.cov	untyped			-

coef.trend	untyped			-

coef.var	untyped			-

control	untyped			-

cov.compute	logical	TRUE	TRUE, FALSE	-

covtype	character	matern5_2	gauss, matern5_2, matern3_2, exp, powexp	-

estim.method	character	MLE	MLE, LOO	-

gr	logical	TRUE	TRUE, FALSE	-

iso	logical	FALSE	TRUE, FALSE	-

jitter	numeric	0		\([0, \infty)\)

kernel	untyped			-

knots	untyped			-

light.return	logical	FALSE	TRUE, FALSE	-

lower	untyped			-

multistart	integer	1		\((-\infty, \infty)\)

noise.var	untyped			-

nugget	numeric	-		\((-\infty, \infty)\)

nugget.estim	logical	FALSE	TRUE, FALSE	-

nugget.stability	numeric	0		\([0, \infty)\)

optim.method	character	BFGS	BFGS, gen	-

parinit	untyped			-

penalty	untyped			-

scaling	logical	FALSE	TRUE, FALSE	-

se.compute	logical	TRUE	TRUE, FALSE	-

type	character	SK	SK, UK	-

upper	untyped			-

References

Roustant O, Ginsbourger D, Deville Y (2012). “DiceKriging, DiceOptim: Two R Packages for the Analysis of Computer Experiments by Kriging-Based Metamodeling and Optimization.” Journal of Statistical Software, 51(1), 1--55. doi:10.18637/jss.v051.i01 .

See also

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrKM

Methods

Public methods

LearnerRegrKM$new()

LearnerRegrKM$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrKM$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrKM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("DiceKriging", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("regr.km")
print(learner)

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerRegrKM:regr.km>: Kriging
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, DiceKriging
#> * Predict Types:  [response], se
#> * Feature Types: logical, integer, numeric
#> * Properties: -
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : BFGS 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern5_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 
#>   - parameters upper bounds :  2 14 8 786.6 2.3 4 538 16.6 2 7.822 
#>   - best initial criterion value(s) :  -60.10881 
#> 
#> N = 10, M = 5 machine precision = 2.22045e-16
#> At X0, 0 variables are exactly at the bounds
#> At iterate     0  f=       60.109  |proj g|=       7.1216
#> At iterate     1  f =       60.099  |proj g|=       0.70696
#> At iterate     2  f =       59.671  |proj g|=       0.58354
#> At iterate     3  f =       59.634  |proj g|=        1.4495
#> At iterate     4  f =        59.61  |proj g|=       0.44288
#> At iterate     5  f =       59.581  |proj g|=       0.37283
#> At iterate     6  f =       59.474  |proj g|=       0.27461
#> At iterate     7  f =       59.398  |proj g|=      0.091069
#> At iterate     8  f =       59.342  |proj g|=       0.04586
#> At iterate     9  f =       59.341  |proj g|=       0.50295
#> At iterate    10  f =       59.338  |proj g|=       0.12785
#> At iterate    11  f =       59.337  |proj g|=       0.04338
#> At iterate    12  f =        59.33  |proj g|=       0.31831
#> At iterate    13  f =       59.321  |proj g|=       0.51443
#> At iterate    14  f =       59.313  |proj g|=       0.42738
#> At iterate    15  f =       59.308  |proj g|=       0.31906
#> At iterate    16  f =       59.303  |proj g|=       0.25188
#> At iterate    17  f =       59.301  |proj g|=       0.25592
#> At iterate    18  f =       59.282  |proj g|=       0.32249
#> At iterate    19  f =        59.28  |proj g|=      0.033639
#> At iterate    20  f =        59.28  |proj g|=     0.0034204
#> At iterate    21  f =        59.28  |proj g|=      0.002311
#> At iterate    22  f =        59.28  |proj g|=     0.0023203
#> At iterate    23  f =        59.28  |proj g|=     0.0058879
#> At iterate    24  f =        59.28  |proj g|=      0.011391
#> At iterate    25  f =        59.28  |proj g|=       0.02091
#> At iterate    26  f =        59.28  |proj g|=      0.033647
#> At iterate    27  f =        59.28  |proj g|=      0.051857
#> At iterate    28  f =        59.28  |proj g|=      0.079729
#> At iterate    29  f =       59.279  |proj g|=       0.12217
#> At iterate    30  f =       59.009  |proj g|=        6.1906
#> ys=-1.266e-01  -gs= 1.300e-02, BFGS update SKIPPED
#> At iterate    31  f =       58.236  |proj g|=        1.8877
#> At iterate    32  f =       57.845  |proj g|=        1.0503
#> At iterate    33  f =       57.488  |proj g|=       0.16872
#> At iterate    34  f =       57.373  |proj g|=       0.24278
#> At iterate    35  f =        57.29  |proj g|=       0.40919
#> At iterate    36  f =       57.221  |proj g|=        0.3089
#> At iterate    37  f =       57.197  |proj g|=       0.13919
#> At iterate    38  f =       57.188  |proj g|=      0.078786
#> At iterate    39  f =       57.187  |proj g|=       0.01067
#> At iterate    40  f =       57.187  |proj g|=       0.00143
#> At iterate    41  f =       57.187  |proj g|=     0.0061044
#> At iterate    42  f =       57.187  |proj g|=      0.016656
#> At iterate    43  f =       57.186  |proj g|=     0.0062119
#> At iterate    44  f =       57.186  |proj g|=    0.00027666
#> At iterate    45  f =       57.186  |proj g|=     0.0033411
#> At iterate    46  f =       57.186  |proj g|=      0.017039
#> At iterate    47  f =       57.181  |proj g|=      0.060357
#> ys=-1.400e-04  -gs= 1.440e-03, BFGS update SKIPPED
#> At iterate    48  f =       57.181  |proj g|=      0.061513
#> At iterate    49  f =       57.179  |proj g|=      0.087088
#> At iterate    50  f =       57.175  |proj g|=       0.14048
#> At iterate    51  f =       57.155  |proj g|=       0.21945
#> At iterate    52  f =       57.128  |proj g|=       0.22139
#> ys=-1.411e-02  -gs= 6.992e-02, BFGS update SKIPPED
#> At iterate    53  f =       57.115  |proj g|=      0.096036
#> At iterate    54  f =       57.091  |proj g|=        0.2542
#> At iterate    55  f =       57.038  |proj g|=       0.55116
#> At iterate    56  f =       57.036  |proj g|=       0.59962
#> At iterate    57  f =       57.003  |proj g|=       0.47298
#> At iterate    58  f =       56.911  |proj g|=       0.28528
#> At iterate    59  f =       56.898  |proj g|=        0.2407
#> At iterate    60  f =       56.854  |proj g|=      0.049152
#> At iterate    61  f =       56.849  |proj g|=      0.028496
#> At iterate    62  f =       56.849  |proj g|=     0.0058821
#> At iterate    63  f =       56.849  |proj g|=    0.00020724
#> At iterate    64  f =       56.849  |proj g|=    0.00020725
#> 
#> iterations 64
#> function evaluations 89
#> segments explored during Cauchy searches 64
#> BFGS updates skipped 3
#> active bounds at final generalized Cauchy point 6
#> norm of the final projected gradient 0.000207246
#> final function value 56.8487
#> 
#> F = 56.8487
#> final  value 56.848664 
#> converged
#> 
#> Call:
#> DiceKriging::km(design = data, response = task$truth(), control = pv$control)
#> 
#> Trend  coeff.:
#>                Estimate
#>  (Intercept)    19.5271
#> 
#> Covar. type  : matern5_2 
#> Covar. coeff.:
#>                Estimate
#>    theta(am)     2.0000
#>  theta(carb)     0.0000
#>   theta(cyl)     8.0000
#>  theta(disp)   373.4357
#>  theta(drat)     2.3000
#>  theta(gear)     0.5334
#>    theta(hp)   475.2210
#>  theta(qsec)     1.2121
#>    theta(vs)     2.0000
#>    theta(wt)     7.8220
#> 
#> Variance estimate: 31.31853
#> regr.mse 
#> 17.42841