Kriging Regression Learner

Kriging regression. Calls DiceKriging::km() from package DiceKriging.

The predict type hyperparameter "type" defaults to "sk" (simple kriging).

The additional hyperparameter nugget.stability is used to overwrite the hyperparameter nugget with nugget.stability * var(y) before training to improve the numerical stability. We recommend a value of 1e-8.

The additional hyperparameter jitter can be set to add N(0, [jitter])-distributed noise to the data before prediction to avoid perfect interpolation. We recommend a value of 1e-12.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.km")
lrn("regr.km")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3learners, DiceKriging

Parameters

Id	Type	Default	Levels	Range

bias.correct	logical	FALSE	TRUE, FALSE	-

checkNames	logical	TRUE	TRUE, FALSE	-

coef.cov	untyped			-

coef.trend	untyped			-

coef.var	untyped			-

control	untyped			-

cov.compute	logical	TRUE	TRUE, FALSE	-

covtype	character	matern5_2	gauss, matern5_2, matern3_2, exp, powexp	-

estim.method	character	MLE	MLE, LOO	-

gr	logical	TRUE	TRUE, FALSE	-

iso	logical	FALSE	TRUE, FALSE	-

jitter	numeric	0		\([0, \infty)\)

kernel	untyped			-

knots	untyped			-

light.return	logical	FALSE	TRUE, FALSE	-

lower	untyped			-

multistart	integer	1		\((-\infty, \infty)\)

noise.var	untyped			-

nugget	numeric	-		\((-\infty, \infty)\)

nugget.estim	logical	FALSE	TRUE, FALSE	-

nugget.stability	numeric	0		\([0, \infty)\)

optim.method	character	BFGS	BFGS, gen	-

parinit	untyped			-

penalty	untyped			-

scaling	logical	FALSE	TRUE, FALSE	-

se.compute	logical	TRUE	TRUE, FALSE	-

type	character	SK	SK, UK	-

upper	untyped			-

References

Roustant O, Ginsbourger D, Deville Y (2012). “DiceKriging, DiceOptim: Two R Packages for the Analysis of Computer Experiments by Kriging-Based Metamodeling and Optimization.” Journal of Statistical Software, 51(1), 1--55. doi:10.18637/jss.v051.i01 .

See also

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrKM

Methods

Public methods

LearnerRegrKM$new()

LearnerRegrKM$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrKM$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrKM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("DiceKriging", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("regr.km")
print(learner)

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerRegrKM:regr.km>: Kriging
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, DiceKriging
#> * Predict Types:  [response], se
#> * Feature Types: logical, integer, numeric
#> * Properties: -
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : BFGS 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern5_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 
#>   - parameters upper bounds :  2 10 8 737.8 4.34 4 424 11.44 2 7.46 
#>   - best initial criterion value(s) :  -59.17765 
#> 
#> N = 10, M = 5 machine precision = 2.22045e-16
#> At X0, 0 variables are exactly at the bounds
#> At iterate     0  f=       59.178  |proj g|=      0.77911
#> At iterate     1  f =       58.039  |proj g|=        0.5291
#> At iterate     2  f =       57.931  |proj g|=        0.1794
#> At iterate     3  f =        57.89  |proj g|=       0.27578
#> At iterate     4  f =       57.807  |proj g|=       0.24461
#> At iterate     5  f =       57.697  |proj g|=       0.14562
#> At iterate     6  f =        57.38  |proj g|=       0.22369
#> At iterate     7  f =       57.276  |proj g|=      0.069001
#> At iterate     8  f =       57.264  |proj g|=        0.0616
#> At iterate     9  f =       57.251  |proj g|=      0.030822
#> At iterate    10  f =       57.248  |proj g|=      0.026988
#> At iterate    11  f =        57.24  |proj g|=      0.044869
#> At iterate    12  f =       57.233  |proj g|=       0.05803
#> At iterate    13  f =       57.213  |proj g|=      0.098047
#> At iterate    14  f =       57.204  |proj g|=       0.11212
#> At iterate    15  f =       57.196  |proj g|=       0.01858
#> At iterate    16  f =       57.194  |proj g|=      0.014429
#> At iterate    17  f =       57.193  |proj g|=      0.016732
#> At iterate    18  f =        57.19  |proj g|=      0.027093
#> At iterate    19  f =       57.183  |proj g|=      0.053006
#> At iterate    20  f =       57.163  |proj g|=       0.12164
#> At iterate    21  f =       57.136  |proj g|=       0.10696
#> At iterate    22  f =       57.117  |proj g|=      0.044492
#> At iterate    23  f =       57.113  |proj g|=      0.019819
#> At iterate    24  f =        57.11  |proj g|=      0.031266
#> At iterate    25  f =       57.106  |proj g|=      0.053922
#> At iterate    26  f =       57.089  |proj g|=       0.10034
#> At iterate    27  f =       57.019  |proj g|=       0.16726
#> At iterate    28  f =       56.996  |proj g|=       0.11608
#> At iterate    29  f =       56.954  |proj g|=      0.035636
#> At iterate    30  f =       56.897  |proj g|=       0.13803
#> At iterate    31  f =       56.861  |proj g|=       0.15753
#> At iterate    32  f =       56.852  |proj g|=       0.05934
#> At iterate    33  f =        56.85  |proj g|=      0.019927
#> At iterate    34  f =        56.85  |proj g|=     0.0066503
#> At iterate    35  f =        56.85  |proj g|=    0.00058827
#> At iterate    36  f =        56.85  |proj g|=    0.00023281
#> At iterate    37  f =        56.85  |proj g|=    0.00023282
#> 
#> iterations 37
#> function evaluations 44
#> segments explored during Cauchy searches 38
#> BFGS updates skipped 0
#> active bounds at final generalized Cauchy point 5
#> norm of the final projected gradient 0.000232823
#> final function value 56.8496
#> 
#> F = 56.8496
#> final  value 56.849616 
#> converged
#> 
#> Call:
#> DiceKriging::km(design = data, response = task$truth(), control = pv$control)
#> 
#> Trend  coeff.:
#>                Estimate
#>  (Intercept)    20.1575
#> 
#> Covar. type  : matern5_2 
#> Covar. coeff.:
#>                Estimate
#>    theta(am)     0.0000
#>  theta(carb)    10.0000
#>   theta(cyl)     8.0000
#>  theta(disp)    71.4296
#>  theta(drat)     1.9795
#>  theta(gear)     4.0000
#>    theta(hp)   411.2459
#>  theta(qsec)     1.8924
#>    theta(vs)     2.0000
#>    theta(wt)     1.7086
#> 
#> Variance estimate: 33.02909
#> regr.mse 
#> 15.72934