Kriging Regression Learner

Kriging regression. Calls DiceKriging::km() from package DiceKriging.

The predict type hyperparameter "type" defaults to "sk" (simple kriging).

The additional hyperparameter nugget.stability is used to overwrite the hyperparameter nugget with nugget.stability * var(y) before training to improve the numerical stability. We recommend a value of 1e-8.

The additional hyperparameter jitter can be set to add N(0, [jitter])-distributed noise to the data before prediction to avoid perfect interpolation. We recommend a value of 1e-12.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("regr.km")
lrn("regr.km")

Meta Information

Task type: “regr”

Predict Types: “response”, “se”

Feature Types: “logical”, “integer”, “numeric”

Required Packages: mlr3, mlr3learners, DiceKriging

Parameters

Id	Type	Default	Levels	Range

bias.correct	logical	FALSE	TRUE, FALSE	-

checkNames	logical	TRUE	TRUE, FALSE	-

coef.cov	untyped			-

coef.trend	untyped			-

coef.var	untyped			-

control	untyped			-

cov.compute	logical	TRUE	TRUE, FALSE	-

covtype	character	matern5_2	gauss, matern5_2, matern3_2, exp, powexp	-

estim.method	character	MLE	MLE, LOO	-

gr	logical	TRUE	TRUE, FALSE	-

iso	logical	FALSE	TRUE, FALSE	-

jitter	numeric	0		\([0, \infty)\)

kernel	untyped			-

knots	untyped			-

light.return	logical	FALSE	TRUE, FALSE	-

lower	untyped			-

multistart	integer	1		\((-\infty, \infty)\)

noise.var	untyped			-

nugget	numeric	-		\((-\infty, \infty)\)

nugget.estim	logical	FALSE	TRUE, FALSE	-

nugget.stability	numeric	0		\([0, \infty)\)

optim.method	character	BFGS	BFGS, gen	-

parinit	untyped			-

penalty	untyped			-

scaling	logical	FALSE	TRUE, FALSE	-

se.compute	logical	TRUE	TRUE, FALSE	-

type	character	SK	SK, UK	-

upper	untyped			-

References

Roustant O, Ginsbourger D, Deville Y (2012). “DiceKriging, DiceOptim: Two R Packages for the Analysis of Computer Experiments by Kriging-Based Metamodeling and Optimization.” Journal of Statistical Software, 51(1), 1--55. doi:10.18637/jss.v051.i01 .

See also

Super classes

mlr3::Learner -> mlr3::LearnerRegr -> LearnerRegrKM

Methods

Public methods

LearnerRegrKM$new()

LearnerRegrKM$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerRegrKM$new()

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerRegrKM$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("DiceKriging", quietly = TRUE)) {
# Define the Learner and set parameter values
learner = lrn("regr.km")
print(learner)

# Define a Task
task = tsk("mtcars")

# Create train and test set
ids = partition(task)

# Train the learner on the training ids
learner$train(task, row_ids = ids$train)

# print the model
print(learner$model)

# importance method
if("importance" %in% learner$properties) print(learner$importance)

# Make predictions for the test rows
predictions = learner$predict(task, row_ids = ids$test)

# Score the predictions
predictions$score()
}
#> <LearnerRegrKM:regr.km>: Kriging
#> * Model: -
#> * Parameters: list()
#> * Packages: mlr3, mlr3learners, DiceKriging
#> * Predict Types:  [response], se
#> * Feature Types: logical, integer, numeric
#> * Properties: -
#> 
#> optimisation start
#> ------------------
#> * estimation method   : MLE 
#> * optimisation method : BFGS 
#> * analytical gradient : used
#> * trend model : ~1
#> * covariance model : 
#>   - type :  matern5_2 
#>   - nugget : NO
#>   - parameters lower bounds :  1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 1e-10 
#>   - parameters upper bounds :  2 14 8 762.6 2.86 4 546 11.02 2 7.822 
#>   - best initial criterion value(s) :  -61.46448 
#> 
#> N = 10, M = 5 machine precision = 2.22045e-16
#> At X0, 0 variables are exactly at the bounds
#> At iterate     0  f=       61.464  |proj g|=       2.4358
#> At iterate     1  f =       57.593  |proj g|=        1.0183
#> At iterate     2  f =       57.457  |proj g|=       0.26249
#> At iterate     3  f =       57.378  |proj g|=       0.41353
#> At iterate     4  f =       57.194  |proj g|=       0.65445
#> At iterate     5  f =       57.112  |proj g|=       0.24714
#> At iterate     6  f =       57.068  |proj g|=         0.228
#> At iterate     7  f =       57.027  |proj g|=       0.20737
#> At iterate     8  f =       56.599  |proj g|=       0.20035
#> At iterate     9  f =       56.347  |proj g|=       0.60328
#> At iterate    10  f =       56.278  |proj g|=       0.33115
#> At iterate    11  f =       56.198  |proj g|=       0.17609
#> At iterate    12  f =       56.054  |proj g|=       0.10897
#> At iterate    13  f =       56.045  |proj g|=      0.048624
#> At iterate    14  f =       56.032  |proj g|=      0.091226
#> At iterate    15  f =       56.001  |proj g|=       0.29493
#> At iterate    16  f =       55.955  |proj g|=       0.34865
#> At iterate    17  f =       55.941  |proj g|=       0.10482
#> At iterate    18  f =       55.933  |proj g|=      0.089522
#> At iterate    19  f =       55.928  |proj g|=       0.06376
#> At iterate    20  f =       55.927  |proj g|=      0.016646
#> At iterate    21  f =       55.927  |proj g|=      0.030683
#> At iterate    22  f =       55.926  |proj g|=      0.072248
#> At iterate    23  f =       55.923  |proj g|=       0.13376
#> At iterate    24  f =       55.917  |proj g|=       0.19673
#> At iterate    25  f =       55.909  |proj g|=       0.20003
#> At iterate    26  f =       55.906  |proj g|=      0.066127
#> At iterate    27  f =       55.904  |proj g|=      0.046649
#> At iterate    28  f =       55.901  |proj g|=        0.0126
#> At iterate    29  f =       55.901  |proj g|=     0.0073084
#> At iterate    30  f =       55.895  |proj g|=      0.041583
#> At iterate    31  f =       55.885  |proj g|=      0.080316
#> At iterate    32  f =       55.879  |proj g|=       0.16896
#> At iterate    33  f =       55.864  |proj g|=      0.095786
#> At iterate    34  f =       55.855  |proj g|=      0.041882
#> At iterate    35  f =       55.854  |proj g|=     0.0066167
#> At iterate    36  f =       55.854  |proj g|=    0.00061138
#> At iterate    37  f =       55.854  |proj g|=     0.0002271
#> 
#> iterations 37
#> function evaluations 45
#> segments explored during Cauchy searches 38
#> BFGS updates skipped 0
#> active bounds at final generalized Cauchy point 6
#> norm of the final projected gradient 0.000227104
#> final function value 55.8538
#> 
#> F = 55.8538
#> final  value 55.853811 
#> converged
#> 
#> Call:
#> DiceKriging::km(design = data, response = task$truth(), control = pv$control)
#> 
#> Trend  coeff.:
#>                Estimate
#>  (Intercept)    19.5905
#> 
#> Covar. type  : matern5_2 
#> Covar. coeff.:
#>                Estimate
#>    theta(am)     2.0000
#>  theta(carb)    14.0000
#>   theta(cyl)     8.0000
#>  theta(disp)    92.9202
#>  theta(drat)     2.8600
#>  theta(gear)     1.2442
#>    theta(hp)   526.9554
#>  theta(qsec)     1.0969
#>    theta(vs)     2.0000
#>    theta(wt)     7.8220
#> 
#> Variance estimate: 28.34269
#> regr.mse 
#> 7.635778