Extreme Gradient Boosting Survival Learner

eXtreme Gradient Boosting regression. Calls xgboost::xgb.train() from package xgboost.

Custom mlr3 defaults

nrounds:

Actual default: no default.

Adjusted default: 1.

Reason for change: Without a default construction of the learner would error. Just setting a nonsense default to workaround this. nrounds needs to be tuned by the user.

nthread:

Actual value: Undefined, triggering auto-detection of the number of CPUs.

Adjusted value: 1.

Reason for change: Conflicting with parallelization via future.

verbose:

Actual default: 1.

Adjusted default: 0.

Reason for change: Reduce verbosity.

objective:

Actual default: reg:squarederror

Adjusted default: survival:cox

Reason for change: Changed to a survival objective.

Dictionary

This Learner can be instantiated via the dictionary mlr_learners or with the associated sugar function lrn():

mlr_learners$get("surv.xgboost")
lrn("surv.xgboost")

Meta Information

Task type: “surv”

Predict Types: “crank”, “lp”

Feature Types: “integer”, “numeric”

Required Packages: xgboost

Parameters

Id	Type	Default	Levels	Range

aft_loss_distribution	character	normal	normal, logistic, extreme	\((-\infty, \infty)\)

aft_loss_distribution_scale	numeric	-		\((-\infty, \infty)\)

alpha	numeric	0		\([0, \infty)\)

base_score	numeric	0.5		\((-\infty, \infty)\)

booster	character	gbtree	gbtree, gblinear, dart	\((-\infty, \infty)\)

callbacks	list	NULL		\((-\infty, \infty)\)

colsample_bylevel	numeric	1		\([0, 1]\)

colsample_bynode	numeric	1		\([0, 1]\)

colsample_bytree	numeric	1		\([0, 1]\)

disable_default_eval_metric	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

early_stopping_rounds	integer	NULL		\([1, \infty)\)

eta	numeric	0.3		\([0, 1]\)

feature_selector	character	cyclic	cyclic, shuffle, random, greedy, thrifty	\((-\infty, \infty)\)

feval	list	NULL		\((-\infty, \infty)\)

gamma	numeric	0		\([0, \infty)\)

grow_policy	character	depthwise	depthwise, lossguide	\((-\infty, \infty)\)

interaction_constraints	list	-		\((-\infty, \infty)\)

lambda	numeric	1		\([0, \infty)\)

lambda_bias	numeric	0		\([0, \infty)\)

max_bin	integer	256		\([2, \infty)\)

max_delta_step	numeric	0		\([0, \infty)\)

max_depth	integer	6		\([0, \infty)\)

max_leaves	integer	0		\([0, \infty)\)

maximize	logical	NULL	TRUE, FALSE	\((-\infty, \infty)\)

min_child_weight	numeric	1		\([0, \infty)\)

missing	numeric	NA		\((-\infty, \infty)\)

monotone_constraints	integer	0		\([-1, 1]\)

normalize_type	character	tree	tree, forest	\((-\infty, \infty)\)

nrounds	integer	-		\([1, \infty)\)

nthread	integer	1		\([1, \infty)\)

ntreelimit	integer	-		\([1, \infty)\)

num_parallel_tree	integer	1		\([1, \infty)\)

objective	character	survival:cox	survival:cox, survival:aft	\((-\infty, \infty)\)

one_drop	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

predictor	character	cpu_predictor	cpu_predictor, gpu_predictor	\((-\infty, \infty)\)

print_every_n	integer	1		\([1, \infty)\)

process_type	character	default	default, update	\((-\infty, \infty)\)

rate_drop	numeric	0		\([0, 1]\)

refresh_leaf	logical	TRUE	TRUE, FALSE	\((-\infty, \infty)\)

sampling_method	character	uniform	uniform, gradient_based	\((-\infty, \infty)\)

sample_type	character	uniform	uniform, weighted	\((-\infty, \infty)\)

save_name	list	-		\((-\infty, \infty)\)

save_period	integer	-		\([0, \infty)\)

scale_pos_weight	numeric	1		\((-\infty, \infty)\)

seed_per_iteration	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

sketch_eps	numeric	0.03		\([0, 1]\)

skip_drop	numeric	0		\([0, 1]\)

single_precision_histogram	logical	FALSE	TRUE, FALSE	\((-\infty, \infty)\)

subsample	numeric	1		\([0, 1]\)

top_k	integer	0		\([0, \infty)\)

tree_method	character	auto	auto, exact, approx, hist, gpu_hist	\((-\infty, \infty)\)

tweedie_variance_power	numeric	1.5		\([1, 2]\)

updater	list	-		\((-\infty, \infty)\)

verbose	integer	1		\([0, 2]\)

watchlist	list	NULL		\((-\infty, \infty)\)

xgb_model	list	-		\((-\infty, \infty)\)

References

Chen, Tianqi, Guestrin, Carlos (2016). “Xgboost: A scalable tree boosting system.” In Proceedings of the 22nd ACM SIGKDD Conference on Knowledge Discovery and Data Mining, 785--794. ACM. doi: 10.1145/2939672.2939785 .

See also

Super classes

mlr3::Learner -> mlr3proba::LearnerSurv -> LearnerSurvXgboost

Methods

Public methods

LearnerSurvXgboost$new()

LearnerSurvXgboost$importance()

LearnerSurvXgboost$clone()

Inherited methods

mlr3::Learner$base_learner()

mlr3::Learner$format()

mlr3::Learner$help()

mlr3::Learner$predict()

mlr3::Learner$predict_newdata()

mlr3::Learner$print()

mlr3::Learner$reset()

mlr3::Learner$train()

Method new()

Creates a new instance of this R6 class.

Usage

LearnerSurvXgboost$new()

Method importance()

The importance scores are calculated with xgboost::xgb.importance().

Usage

LearnerSurvXgboost$importance()

Returns

Named numeric().

Method clone()

The objects of this class are cloneable with this method.

Usage

LearnerSurvXgboost$clone(deep = FALSE)

Arguments

deep

Whether to make a deep clone.

Examples

if (requireNamespace("xgboost")) {
  learner = mlr3::lrn("surv.xgboost")
  print(learner)

  # available parameters:
learner$param_set$ids()
}
#> <LearnerSurvXgboost:surv.xgboost>
#> * Model: -
#> * Parameters: nrounds=1, nthread=1, verbose=0
#> * Packages: xgboost
#> * Predict Type: crank
#> * Feature types: integer, numeric
#> * Properties: importance, missings, weights
#>  [1] "aft_loss_distribution"       "aft_loss_distribution_scale"
#>  [3] "alpha"                       "base_score"                 
#>  [5] "booster"                     "callbacks"                  
#>  [7] "colsample_bylevel"           "colsample_bynode"           
#>  [9] "colsample_bytree"            "disable_default_eval_metric"
#> [11] "early_stopping_rounds"       "eta"                        
#> [13] "feature_selector"            "feval"                      
#> [15] "gamma"                       "grow_policy"                
#> [17] "interaction_constraints"     "lambda"                     
#> [19] "lambda_bias"                 "max_bin"                    
#> [21] "max_delta_step"              "max_depth"                  
#> [23] "max_leaves"                  "maximize"                   
#> [25] "min_child_weight"            "missing"                    
#> [27] "monotone_constraints"        "normalize_type"             
#> [29] "nrounds"                     "nthread"                    
#> [31] "ntreelimit"                  "num_parallel_tree"          
#> [33] "objective"                   "one_drop"                   
#> [35] "predictor"                   "print_every_n"              
#> [37] "process_type"                "rate_drop"                  
#> [39] "refresh_leaf"                "sampling_method"            
#> [41] "sample_type"                 "save_name"                  
#> [43] "save_period"                 "scale_pos_weight"           
#> [45] "seed_per_iteration"          "sketch_eps"                 
#> [47] "skip_drop"                   "single_precision_histogram" 
#> [49] "subsample"                   "top_k"                      
#> [51] "tree_method"                 "tweedie_variance_power"     
#> [53] "updater"                     "verbose"                    
#> [55] "watchlist"                   "xgb_model"